update radio docs
