+++ packages/govtnz-ds-website/src/commons/examples/components__Radios.ts
@@ -18,81 +18,6 @@ export default [
     "angular": ""
   },
   {
-    "mustache": "// Developer note: ensure your \".mustache\" files are imported as plain text. In Webpack you might use https://github.com/webpack-contrib/raw-loader\nimport Mustache from \"mustache\";\nimport FieldsetBlock from \"@govtnz/ds/build/mustache/FieldsetBlock.mustache\";\nimport H1 from \"@govtnz/ds/build/mustache/H1.mustache\";\nimport Radios from \"@govtnz/ds/build/mustache/Radios.mustache\";\nimport RadioBlock from \"@govtnz/ds/build/mustache/RadioBlock.mustache\";\n\n// The variables FieldsetBlock, H1, Radios, RadioBlock are all strings that are mustache templates.\n\nexport default `${Mustache.render(FieldsetBlock, {\n  legend: `${Mustache.render(H1, {\n    \"styleSize=large\": true,\n    styleSize: true,\n    id: `nameChangeId5`,\n    children: `Have you changed your name?`\n  })}`,\n  hint: `<p >This includes changing your last name or spelling your name differently.</p>`,\n  hintId: `hintId5`,\n  errorId: true,\n  error: `Select 'yes' if you have changed your name.`,\n  children: `${Mustache.render(Radios, {\n    inline: true,\n    children: `${Mustache.render(RadioBlock, {\n      label: `Yes`,\n      id: `anyRadioId9234`,\n      value: `true`,\n      name: `nameChange5`,\n      labelId: `labelId9`\n    })}${Mustache.render(RadioBlock, {\n      label: `No`,\n      id: `anyRadioId10`,\n      value: `false`,\n      name: `nameChange5`,\n      labelId: `labelId10`\n    })}`\n  })}`\n})}`;\n",
-    "silverstripe-components": "",
-    "react-js": "import React from \"react\";\nimport FieldsetBlock from \"@govtnz/ds/build/react-js/FieldsetBlock\";\nimport \"@govtnz/ds/build/css/FieldsetBlock.css\"; // or @govtnz/ds/build/scss/FieldsetBlock.scss\nimport H1 from \"@govtnz/ds/build/react-js/H1\";\nimport \"@govtnz/ds/build/css/H1.css\"; // or @govtnz/ds/build/scss/H1.scss\nimport Radios from \"@govtnz/ds/build/react-js/Radios\";\nimport \"@govtnz/ds/build/css/Radios.css\"; // or @govtnz/ds/build/scss/Radios.scss\nimport RadioBlock from \"@govtnz/ds/build/react-js/RadioBlock\";\nimport \"@govtnz/ds/build/css/RadioBlock.css\"; // or @govtnz/ds/build/scss/RadioBlock.scss\n\nexport default () => (\n  <FieldsetBlock\n    legend={\n      <H1 styleSize=\"large\" id=\"nameChangeId5\">\n        Have you changed your name?\n      </H1>\n    }\n    hint={\n      <p>\n        This includes changing your last name or spelling your name differently.\n      </p>\n    }\n    hintId=\"hintId5\"\n    errorId=\"errorId5\"\n    error=\"Select 'yes' if you have changed your name.\"\n  >\n    <Radios inline>\n      <RadioBlock\n        label=\"Yes\"\n        id=\"anyRadioId9234\"\n        value=\"true\"\n        name=\"nameChange5\"\n        labelId=\"labelId9\"\n      />\n      <RadioBlock\n        label=\"No\"\n        id=\"anyRadioId10\"\n        value=\"false\"\n        name=\"nameChange5\"\n        labelId=\"labelId10\"\n      />\n    </Radios>\n  </FieldsetBlock>\n);\n",
-    "react-ts": "import React from \"react\";\nimport FieldsetBlock from \"@govtnz/ds/build/react-ts/FieldsetBlock\";\nimport \"@govtnz/ds/build/css/FieldsetBlock.css\"; // or @govtnz/ds/build/scss/FieldsetBlock.scss\nimport H1 from \"@govtnz/ds/build/react-ts/H1\";\nimport \"@govtnz/ds/build/css/H1.css\"; // or @govtnz/ds/build/scss/H1.scss\nimport Radios from \"@govtnz/ds/build/react-ts/Radios\";\nimport \"@govtnz/ds/build/css/Radios.css\"; // or @govtnz/ds/build/scss/Radios.scss\nimport RadioBlock from \"@govtnz/ds/build/react-ts/RadioBlock\";\nimport \"@govtnz/ds/build/css/RadioBlock.css\"; // or @govtnz/ds/build/scss/RadioBlock.scss\n\nexport default () => (\n  <FieldsetBlock\n    legend={\n      <H1 styleSize=\"large\" id=\"nameChangeId5\">\n        Have you changed your name?\n      </H1>\n    }\n    hint={\n      <p>\n        This includes changing your last name or spelling your name differently.\n      </p>\n    }\n    hintId=\"hintId5\"\n    errorId=\"errorId5\"\n    error=\"Select 'yes' if you have changed your name.\"\n  >\n    <Radios inline>\n      <RadioBlock\n        label=\"Yes\"\n        id=\"anyRadioId9234\"\n        value=\"true\"\n        name=\"nameChange5\"\n        labelId=\"labelId9\"\n      />\n      <RadioBlock\n        label=\"No\"\n        id=\"anyRadioId10\"\n        value=\"false\"\n        name=\"nameChange5\"\n        labelId=\"labelId10\"\n      />\n    </Radios>\n  </FieldsetBlock>\n);\n",
-    "react-js-styled-components": "import React from \"react\";\nimport FieldsetBlock from \"@govtnz/ds/build/react-js-styled-components/FieldsetBlock\";\nimport H1 from \"@govtnz/ds/build/react-js-styled-components/H1\";\nimport Radios from \"@govtnz/ds/build/react-js-styled-components/Radios\";\nimport RadioBlock from \"@govtnz/ds/build/react-js-styled-components/RadioBlock\";\n\nexport default () => (\n  <FieldsetBlock\n    legend={\n      <H1 styleSize=\"large\" id=\"nameChangeId5\">\n        Have you changed your name?\n      </H1>\n    }\n    hint={\n      <p>\n        This includes changing your last name or spelling your name differently.\n      </p>\n    }\n    hintId=\"hintId5\"\n    errorId=\"errorId5\"\n    error=\"Select 'yes' if you have changed your name.\"\n  >\n    <Radios inline>\n      <RadioBlock\n        label=\"Yes\"\n        id=\"anyRadioId9234\"\n        value=\"true\"\n        name=\"nameChange5\"\n        labelId=\"labelId9\"\n      />\n      <RadioBlock\n        label=\"No\"\n        id=\"anyRadioId10\"\n        value=\"false\"\n        name=\"nameChange5\"\n        labelId=\"labelId10\"\n      />\n    </Radios>\n  </FieldsetBlock>\n);\n",
-    "react-ts-styled-components": "import React from \"react\";\nimport FieldsetBlock from \"@govtnz/ds/build/react-ts-styled-components/FieldsetBlock\";\nimport H1 from \"@govtnz/ds/build/react-ts-styled-components/H1\";\nimport Radios from \"@govtnz/ds/build/react-ts-styled-components/Radios\";\nimport RadioBlock from \"@govtnz/ds/build/react-ts-styled-components/RadioBlock\";\n\nexport default () => (\n  <FieldsetBlock\n    legend={\n      <H1 styleSize=\"large\" id=\"nameChangeId5\">\n        Have you changed your name?\n      </H1>\n    }\n    hint={\n      <p>\n        This includes changing your last name or spelling your name differently.\n      </p>\n    }\n    hintId=\"hintId5\"\n    errorId=\"errorId5\"\n    error=\"Select 'yes' if you have changed your name.\"\n  >\n    <Radios inline>\n      <RadioBlock\n        label=\"Yes\"\n        id=\"anyRadioId9234\"\n        value=\"true\"\n        name=\"nameChange5\"\n        labelId=\"labelId9\"\n      />\n      <RadioBlock\n        label=\"No\"\n        id=\"anyRadioId10\"\n        value=\"false\"\n        name=\"nameChange5\"\n        labelId=\"labelId10\"\n      />\n    </Radios>\n  </FieldsetBlock>\n);\n",
-    "vue-js": "<template>\n  <fieldset-block\n    legend=\"Have you changed your name?\"\n    hint=\"This includes changing your last name or spelling your name differently.\"\n    hintId=\"hintId5\"\n    errorId=\"errorId5\"\n    error=\"Select 'yes' if you have changed your name.\"\n  >\n    <c-radios inline>\n      <radio-block\n        label=\"Yes\"\n        id=\"anyRadioId9234\"\n        value=\"true\"\n        name=\"nameChange5\"\n        labelId=\"labelId9\"\n      />\n      <radio-block\n        label=\"No\"\n        id=\"anyRadioId10\"\n        value=\"false\"\n        name=\"nameChange5\"\n        labelId=\"labelId10\"\n      />\n    </c-radios>\n  </fieldset-block>\n</template>\n<script>\nimport Vue from \"vue\";\nimport FieldsetBlock from \"@govtnz/ds/build/vue-js/FieldsetBlock.vue\";\nimport H1 from \"@govtnz/ds/build/vue-js/H1.vue\";\nimport Radios from \"@govtnz/ds/build/vue-js/Radios.vue\";\nimport RadioBlock from \"@govtnz/ds/build/vue-js/RadioBlock.vue\";\n\nexport default {\n  components: {\n    \"fieldset-block\": FieldsetBlock,\n    \"c-h1\": H1,\n    \"c-radios\": Radios,\n    \"radio-block\": RadioBlock\n  }\n};\n</script>\n",
-    "vue-ts": "<template>\n  <fieldset-block\n    legend=\"Have you changed your name?\"\n    hint=\"This includes changing your last name or spelling your name differently.\"\n    hintId=\"hintId5\"\n    errorId=\"errorId5\"\n    error=\"Select 'yes' if you have changed your name.\"\n  >\n    <c-radios inline>\n      <radio-block\n        label=\"Yes\"\n        id=\"anyRadioId9234\"\n        value=\"true\"\n        name=\"nameChange5\"\n        labelId=\"labelId9\"\n      />\n      <radio-block\n        label=\"No\"\n        id=\"anyRadioId10\"\n        value=\"false\"\n        name=\"nameChange5\"\n        labelId=\"labelId10\"\n      />\n    </c-radios>\n  </fieldset-block>\n</template>\n<script lang=\"ts\">\nimport Vue from \"vue\";\nimport FieldsetBlock from \"@govtnz/ds/build/vue-ts/FieldsetBlock.vue\";\nimport H1 from \"@govtnz/ds/build/vue-ts/H1.vue\";\nimport Radios from \"@govtnz/ds/build/vue-ts/Radios.vue\";\nimport RadioBlock from \"@govtnz/ds/build/vue-ts/RadioBlock.vue\";\n\nexport default {\n  components: {\n    \"fieldset-block\": FieldsetBlock,\n    \"c-h1\": H1,\n    \"c-radios\": Radios,\n    \"radio-block\": RadioBlock\n  }\n};\n</script>\n",
-    "scss": "",
-    "html": "<!--\nRemember to add these styles:\nin CSS: FieldsetBlock.css, H1.css, Radios.css, RadioBlock.css\nOR in Sass (SCSS): FieldsetBlock.scss, H1.scss, Radios.scss, RadioBlock.scss\n-->\n<div class=\"g-fieldsetBlock-form-group\">\n  <fieldset\n    class=\"g-fieldsetBlock-fieldset\"\n    aria-describedby=\"hintId5 errorId5\"\n  >\n    <legend class=\"g-fieldsetBlock-fieldset__legend\">\n      <h1 class=\"g-heading g-heading-l\" id=\"nameChangeId5\">\n        Have you changed your name?\n      </h1>\n    </legend>\n    <div class=\"g-fieldsetBlock-hint\" id=\"hintId5\">\n      <p>\n        This includes changing your last name or spelling your name differently.\n      </p>\n    </div>\n    <div class=\"g-fieldsetBlock-error-message\" id=\"errorId5\">\n      <span class=\"g-fieldsetBlock-visually-hidden\"> Error: </span> Select 'yes'\n      if you have changed your name.\n    </div>\n    <div>\n      <div class=\"g-radios g-radios--inline\">\n        <div class=\"g-radios__item\">\n          <input\n            class=\"g-radios__input\"\n            id=\"anyRadioId9234\"\n            name=\"nameChange5\"\n            type=\"radio\"\n            value=\"true\"\n          />\n          <label class=\"g-radios__label\" for=\"anyRadioId9234\"> Yes </label>\n        </div>\n        <div class=\"g-radios__item\">\n          <input\n            class=\"g-radios__input\"\n            id=\"anyRadioId10\"\n            name=\"nameChange5\"\n            type=\"radio\"\n            value=\"false\"\n          />\n          <label class=\"g-radios__label\" for=\"anyRadioId10\"> No </label>\n        </div>\n      </div>\n    </div>\n  </fieldset>\n</div>\n",
-    "css": "",
-    "twig-embed": "{% embed \"FieldsetBlock.html.twig\"  with {'hintId':'hintId5', 'errorId':'errorId5', 'error':'Select 'yes' if you have changed your name.'} only %}{% block legend %}{% embed \"H1.html.twig\"  with {'styleSize':'large', 'id':'nameChangeId5', 'children':'Have you changed your name?'} only %}{% endembed %}{% endblock %}{% block hint %}<p >This includes changing your last name or spelling your name differently.</p>{% endblock %}{% block children %}{% embed \"Radios.html.twig\"  with {'inline':'true'} only %}{% block children %}{% embed \"RadioBlock.html.twig\"  with {'label':'Yes', 'id':'anyRadioId9234', 'value':'true', 'name':'nameChange5', 'labelId':'labelId9'} only %}{% endembed %}\n{% embed \"RadioBlock.html.twig\"  with {'label':'No', 'id':'anyRadioId10', 'value':'false', 'name':'nameChange5', 'labelId':'labelId10'} only %}{% endembed %}{% endblock %}{% endembed %}{% endblock %}{% endembed %}",
-    "angular": ""
-  },
-  {
-    "mustache": "// Developer note: ensure your \".mustache\" files are imported as plain text. In Webpack you might use https://github.com/webpack-contrib/raw-loader\nimport Mustache from \"mustache\";\nimport FieldsetBlock from \"@govtnz/ds/build/mustache/FieldsetBlock.mustache\";\nimport H1 from \"@govtnz/ds/build/mustache/H1.mustache\";\nimport Radios from \"@govtnz/ds/build/mustache/Radios.mustache\";\nimport RadioBlock from \"@govtnz/ds/build/mustache/RadioBlock.mustache\";\n\n// The variables FieldsetBlock, H1, Radios, RadioBlock are all strings that are mustache templates.\n\nexport default `${Mustache.render(FieldsetBlock, {\n  legend: `${Mustache.render(H1, {\n    \"styleSize=large\": true,\n    styleSize: true,\n    id: `nameChangeId1`,\n    children: `Have you changed your name?`\n  })}`,\n  hint: `<p >This includes changing your last name or spelling your name differently.</p>`,\n  hintId: `hintId1`,\n  children: `${Mustache.render(Radios, {\n    inline: true,\n    big: ``,\n    children: `${Mustache.render(RadioBlock, {\n      label: `Yes`,\n      id: `anyRadioId`,\n      value: `true`,\n      name: `nameChange1`,\n      labelId: `labelId1`\n    })}${Mustache.render(RadioBlock, {\n      label: `No`,\n      id: `anyRadioId255`,\n      value: `false`,\n      name: `nameChange1`,\n      labelId: `labelId255`\n    })}`\n  })}`\n})}`;\n",
-    "silverstripe-components": "",
-    "react-js": "import React from \"react\";\nimport FieldsetBlock from \"@govtnz/ds/build/react-js/FieldsetBlock\";\nimport \"@govtnz/ds/build/css/FieldsetBlock.css\"; // or @govtnz/ds/build/scss/FieldsetBlock.scss\nimport H1 from \"@govtnz/ds/build/react-js/H1\";\nimport \"@govtnz/ds/build/css/H1.css\"; // or @govtnz/ds/build/scss/H1.scss\nimport Radios from \"@govtnz/ds/build/react-js/Radios\";\nimport \"@govtnz/ds/build/css/Radios.css\"; // or @govtnz/ds/build/scss/Radios.scss\nimport RadioBlock from \"@govtnz/ds/build/react-js/RadioBlock\";\nimport \"@govtnz/ds/build/css/RadioBlock.css\"; // or @govtnz/ds/build/scss/RadioBlock.scss\n\nexport default () => (\n  <FieldsetBlock\n    legend={\n      <H1 styleSize=\"large\" id=\"nameChangeId1\">\n        Have you changed your name?\n      </H1>\n    }\n    hint={\n      <p>\n        This includes changing your last name or spelling your name differently.\n      </p>\n    }\n    hintId=\"hintId1\"\n  >\n    <Radios inline big>\n      <RadioBlock\n        label=\"Yes\"\n        id=\"anyRadioId\"\n        value=\"true\"\n        name=\"nameChange1\"\n        labelId=\"labelId1\"\n      />\n      <RadioBlock\n        label=\"No\"\n        id=\"anyRadioId255\"\n        value=\"false\"\n        name=\"nameChange1\"\n        labelId=\"labelId255\"\n      />\n    </Radios>\n  </FieldsetBlock>\n);\n",
-    "react-ts": "import React from \"react\";\nimport FieldsetBlock from \"@govtnz/ds/build/react-ts/FieldsetBlock\";\nimport \"@govtnz/ds/build/css/FieldsetBlock.css\"; // or @govtnz/ds/build/scss/FieldsetBlock.scss\nimport H1 from \"@govtnz/ds/build/react-ts/H1\";\nimport \"@govtnz/ds/build/css/H1.css\"; // or @govtnz/ds/build/scss/H1.scss\nimport Radios from \"@govtnz/ds/build/react-ts/Radios\";\nimport \"@govtnz/ds/build/css/Radios.css\"; // or @govtnz/ds/build/scss/Radios.scss\nimport RadioBlock from \"@govtnz/ds/build/react-ts/RadioBlock\";\nimport \"@govtnz/ds/build/css/RadioBlock.css\"; // or @govtnz/ds/build/scss/RadioBlock.scss\n\nexport default () => (\n  <FieldsetBlock\n    legend={\n      <H1 styleSize=\"large\" id=\"nameChangeId1\">\n        Have you changed your name?\n      </H1>\n    }\n    hint={\n      <p>\n        This includes changing your last name or spelling your name differently.\n      </p>\n    }\n    hintId=\"hintId1\"\n  >\n    <Radios inline big>\n      <RadioBlock\n        label=\"Yes\"\n        id=\"anyRadioId\"\n        value=\"true\"\n        name=\"nameChange1\"\n        labelId=\"labelId1\"\n      />\n      <RadioBlock\n        label=\"No\"\n        id=\"anyRadioId255\"\n        value=\"false\"\n        name=\"nameChange1\"\n        labelId=\"labelId255\"\n      />\n    </Radios>\n  </FieldsetBlock>\n);\n",
-    "react-js-styled-components": "import React from \"react\";\nimport FieldsetBlock from \"@govtnz/ds/build/react-js-styled-components/FieldsetBlock\";\nimport H1 from \"@govtnz/ds/build/react-js-styled-components/H1\";\nimport Radios from \"@govtnz/ds/build/react-js-styled-components/Radios\";\nimport RadioBlock from \"@govtnz/ds/build/react-js-styled-components/RadioBlock\";\n\nexport default () => (\n  <FieldsetBlock\n    legend={\n      <H1 styleSize=\"large\" id=\"nameChangeId1\">\n        Have you changed your name?\n      </H1>\n    }\n    hint={\n      <p>\n        This includes changing your last name or spelling your name differently.\n      </p>\n    }\n    hintId=\"hintId1\"\n  >\n    <Radios inline big>\n      <RadioBlock\n        label=\"Yes\"\n        id=\"anyRadioId\"\n        value=\"true\"\n        name=\"nameChange1\"\n        labelId=\"labelId1\"\n      />\n      <RadioBlock\n        label=\"No\"\n        id=\"anyRadioId255\"\n        value=\"false\"\n        name=\"nameChange1\"\n        labelId=\"labelId255\"\n      />\n    </Radios>\n  </FieldsetBlock>\n);\n",
-    "react-ts-styled-components": "import React from \"react\";\nimport FieldsetBlock from \"@govtnz/ds/build/react-ts-styled-components/FieldsetBlock\";\nimport H1 from \"@govtnz/ds/build/react-ts-styled-components/H1\";\nimport Radios from \"@govtnz/ds/build/react-ts-styled-components/Radios\";\nimport RadioBlock from \"@govtnz/ds/build/react-ts-styled-components/RadioBlock\";\n\nexport default () => (\n  <FieldsetBlock\n    legend={\n      <H1 styleSize=\"large\" id=\"nameChangeId1\">\n        Have you changed your name?\n      </H1>\n    }\n    hint={\n      <p>\n        This includes changing your last name or spelling your name differently.\n      </p>\n    }\n    hintId=\"hintId1\"\n  >\n    <Radios inline big>\n      <RadioBlock\n        label=\"Yes\"\n        id=\"anyRadioId\"\n        value=\"true\"\n        name=\"nameChange1\"\n        labelId=\"labelId1\"\n      />\n      <RadioBlock\n        label=\"No\"\n        id=\"anyRadioId255\"\n        value=\"false\"\n        name=\"nameChange1\"\n        labelId=\"labelId255\"\n      />\n    </Radios>\n  </FieldsetBlock>\n);\n",
-    "vue-js": "<template>\n  <fieldset-block\n    legend=\"Have you changed your name?\"\n    hint=\"This includes changing your last name or spelling your name differently.\"\n    hintId=\"hintId1\"\n  >\n    <c-radios inline big>\n      <radio-block\n        label=\"Yes\"\n        id=\"anyRadioId\"\n        value=\"true\"\n        name=\"nameChange1\"\n        labelId=\"labelId1\"\n      />\n      <radio-block\n        label=\"No\"\n        id=\"anyRadioId255\"\n        value=\"false\"\n        name=\"nameChange1\"\n        labelId=\"labelId255\"\n      />\n    </c-radios>\n  </fieldset-block>\n</template>\n<script>\nimport Vue from \"vue\";\nimport FieldsetBlock from \"@govtnz/ds/build/vue-js/FieldsetBlock.vue\";\nimport H1 from \"@govtnz/ds/build/vue-js/H1.vue\";\nimport Radios from \"@govtnz/ds/build/vue-js/Radios.vue\";\nimport RadioBlock from \"@govtnz/ds/build/vue-js/RadioBlock.vue\";\n\nexport default {\n  components: {\n    \"fieldset-block\": FieldsetBlock,\n    \"c-h1\": H1,\n    \"c-radios\": Radios,\n    \"radio-block\": RadioBlock\n  }\n};\n</script>\n",
-    "vue-ts": "<template>\n  <fieldset-block\n    legend=\"Have you changed your name?\"\n    hint=\"This includes changing your last name or spelling your name differently.\"\n    hintId=\"hintId1\"\n  >\n    <c-radios inline big>\n      <radio-block\n        label=\"Yes\"\n        id=\"anyRadioId\"\n        value=\"true\"\n        name=\"nameChange1\"\n        labelId=\"labelId1\"\n      />\n      <radio-block\n        label=\"No\"\n        id=\"anyRadioId255\"\n        value=\"false\"\n        name=\"nameChange1\"\n        labelId=\"labelId255\"\n      />\n    </c-radios>\n  </fieldset-block>\n</template>\n<script lang=\"ts\">\nimport Vue from \"vue\";\nimport FieldsetBlock from \"@govtnz/ds/build/vue-ts/FieldsetBlock.vue\";\nimport H1 from \"@govtnz/ds/build/vue-ts/H1.vue\";\nimport Radios from \"@govtnz/ds/build/vue-ts/Radios.vue\";\nimport RadioBlock from \"@govtnz/ds/build/vue-ts/RadioBlock.vue\";\n\nexport default {\n  components: {\n    \"fieldset-block\": FieldsetBlock,\n    \"c-h1\": H1,\n    \"c-radios\": Radios,\n    \"radio-block\": RadioBlock\n  }\n};\n</script>\n",
-    "scss": "",
-    "html": "<!--\nRemember to add these styles:\nin CSS: FieldsetBlock.css, H1.css, Radios.css, RadioBlock.css\nOR in Sass (SCSS): FieldsetBlock.scss, H1.scss, Radios.scss, RadioBlock.scss\n-->\n<div class=\"g-fieldsetBlock-form-group\">\n  <fieldset class=\"g-fieldsetBlock-fieldset\" aria-describedby=\"hintId1\">\n    <legend class=\"g-fieldsetBlock-fieldset__legend\">\n      <h1 class=\"g-heading g-heading-l\" id=\"nameChangeId1\">\n        Have you changed your name?\n      </h1>\n    </legend>\n    <div class=\"g-fieldsetBlock-hint\" id=\"hintId1\">\n      <p>\n        This includes changing your last name or spelling your name differently.\n      </p>\n    </div>\n    <div>\n      <div class=\"g-radios g-radios--inline g-radios--big\">\n        <div class=\"g-radios__item\">\n          <input\n            class=\"g-radios__input\"\n            id=\"anyRadioId\"\n            name=\"nameChange1\"\n            type=\"radio\"\n            value=\"true\"\n          />\n          <label class=\"g-radios__label\" for=\"anyRadioId\"> Yes </label>\n        </div>\n        <div class=\"g-radios__item\">\n          <input\n            class=\"g-radios__input\"\n            id=\"anyRadioId255\"\n            name=\"nameChange1\"\n            type=\"radio\"\n            value=\"false\"\n          />\n          <label class=\"g-radios__label\" for=\"anyRadioId255\"> No </label>\n        </div>\n      </div>\n    </div>\n  </fieldset>\n</div>\n",
-    "css": "",
-    "twig-embed": "{% embed \"FieldsetBlock.html.twig\"  with {'hintId':'hintId1'} only %}{% block legend %}{% embed \"H1.html.twig\"  with {'styleSize':'large', 'id':'nameChangeId1', 'children':'Have you changed your name?'} only %}{% endembed %}{% endblock %}{% block hint %}<p >This includes changing your last name or spelling your name differently.</p>{% endblock %}{% block children %}{% embed \"Radios.html.twig\"  with {'inline':'true', 'big':'true'} only %}{% block children %}{% embed \"RadioBlock.html.twig\"  with {'label':'Yes', 'id':'anyRadioId', 'value':'true', 'name':'nameChange1', 'labelId':'labelId1'} only %}{% endembed %}\n{% embed \"RadioBlock.html.twig\"  with {'label':'No', 'id':'anyRadioId255', 'value':'false', 'name':'nameChange1', 'labelId':'labelId255'} only %}{% endembed %}{% endblock %}{% endembed %}{% endblock %}{% endembed %}",
-    "angular": ""
-  },
-  {
-    "mustache": "// Developer note: ensure your \".mustache\" files are imported as plain text. In Webpack you might use https://github.com/webpack-contrib/raw-loader\nimport Mustache from \"mustache\";\nimport FieldsetBlock from \"@govtnz/ds/build/mustache/FieldsetBlock.mustache\";\nimport H1 from \"@govtnz/ds/build/mustache/H1.mustache\";\nimport Radios from \"@govtnz/ds/build/mustache/Radios.mustache\";\nimport RadioBlock from \"@govtnz/ds/build/mustache/RadioBlock.mustache\";\n\n// The variables FieldsetBlock, H1, Radios, RadioBlock are all strings that are mustache templates.\n\nexport default `${Mustache.render(FieldsetBlock, {\n  legend: `${Mustache.render(H1, {\n    \"styleSize=large\": true,\n    styleSize: true,\n    id: `nameChangeId5`,\n    children: `Have you changed your name?`\n  })}`,\n  hint: `<p >This includes changing your last name or spelling your name differently.</p>`,\n  hintId: `hintId5`,\n  errorId: true,\n  error: `Select 'yes' if you have changed your name.`,\n  children: `${Mustache.render(Radios, {\n    big: ``,\n    children: `${Mustache.render(RadioBlock, {\n      label: `Yes`,\n      id: `anyRadioId9234`,\n      value: `true`,\n      name: `nameChange5`,\n      labelId: `labelId9`\n    })}${Mustache.render(RadioBlock, {\n      label: `No`,\n      id: `anyRadioId10`,\n      value: `false`,\n      name: `nameChange5`,\n      labelId: `labelId10`\n    })}`\n  })}`\n})}`;\n",
-    "silverstripe-components": "",
-    "react-js": "import React from \"react\";\nimport FieldsetBlock from \"@govtnz/ds/build/react-js/FieldsetBlock\";\nimport \"@govtnz/ds/build/css/FieldsetBlock.css\"; // or @govtnz/ds/build/scss/FieldsetBlock.scss\nimport H1 from \"@govtnz/ds/build/react-js/H1\";\nimport \"@govtnz/ds/build/css/H1.css\"; // or @govtnz/ds/build/scss/H1.scss\nimport Radios from \"@govtnz/ds/build/react-js/Radios\";\nimport \"@govtnz/ds/build/css/Radios.css\"; // or @govtnz/ds/build/scss/Radios.scss\nimport RadioBlock from \"@govtnz/ds/build/react-js/RadioBlock\";\nimport \"@govtnz/ds/build/css/RadioBlock.css\"; // or @govtnz/ds/build/scss/RadioBlock.scss\n\nexport default () => (\n  <FieldsetBlock\n    legend={\n      <H1 styleSize=\"large\" id=\"nameChangeId5\">\n        Have you changed your name?\n      </H1>\n    }\n    hint={\n      <p>\n        This includes changing your last name or spelling your name differently.\n      </p>\n    }\n    hintId=\"hintId5\"\n    errorId=\"errorId5\"\n    error=\"Select 'yes' if you have changed your name.\"\n  >\n    <Radios big>\n      <RadioBlock\n        label=\"Yes\"\n        id=\"anyRadioId9234\"\n        value=\"true\"\n        name=\"nameChange5\"\n        labelId=\"labelId9\"\n      />\n      <RadioBlock\n        label=\"No\"\n        id=\"anyRadioId10\"\n        value=\"false\"\n        name=\"nameChange5\"\n        labelId=\"labelId10\"\n      />\n    </Radios>\n  </FieldsetBlock>\n);\n",
-    "react-ts": "import React from \"react\";\nimport FieldsetBlock from \"@govtnz/ds/build/react-ts/FieldsetBlock\";\nimport \"@govtnz/ds/build/css/FieldsetBlock.css\"; // or @govtnz/ds/build/scss/FieldsetBlock.scss\nimport H1 from \"@govtnz/ds/build/react-ts/H1\";\nimport \"@govtnz/ds/build/css/H1.css\"; // or @govtnz/ds/build/scss/H1.scss\nimport Radios from \"@govtnz/ds/build/react-ts/Radios\";\nimport \"@govtnz/ds/build/css/Radios.css\"; // or @govtnz/ds/build/scss/Radios.scss\nimport RadioBlock from \"@govtnz/ds/build/react-ts/RadioBlock\";\nimport \"@govtnz/ds/build/css/RadioBlock.css\"; // or @govtnz/ds/build/scss/RadioBlock.scss\n\nexport default () => (\n  <FieldsetBlock\n    legend={\n      <H1 styleSize=\"large\" id=\"nameChangeId5\">\n        Have you changed your name?\n      </H1>\n    }\n    hint={\n      <p>\n        This includes changing your last name or spelling your name differently.\n      </p>\n    }\n    hintId=\"hintId5\"\n    errorId=\"errorId5\"\n    error=\"Select 'yes' if you have changed your name.\"\n  >\n    <Radios big>\n      <RadioBlock\n        label=\"Yes\"\n        id=\"anyRadioId9234\"\n        value=\"true\"\n        name=\"nameChange5\"\n        labelId=\"labelId9\"\n      />\n      <RadioBlock\n        label=\"No\"\n        id=\"anyRadioId10\"\n        value=\"false\"\n        name=\"nameChange5\"\n        labelId=\"labelId10\"\n      />\n    </Radios>\n  </FieldsetBlock>\n);\n",
-    "react-js-styled-components": "import React from \"react\";\nimport FieldsetBlock from \"@govtnz/ds/build/react-js-styled-components/FieldsetBlock\";\nimport H1 from \"@govtnz/ds/build/react-js-styled-components/H1\";\nimport Radios from \"@govtnz/ds/build/react-js-styled-components/Radios\";\nimport RadioBlock from \"@govtnz/ds/build/react-js-styled-components/RadioBlock\";\n\nexport default () => (\n  <FieldsetBlock\n    legend={\n      <H1 styleSize=\"large\" id=\"nameChangeId5\">\n        Have you changed your name?\n      </H1>\n    }\n    hint={\n      <p>\n        This includes changing your last name or spelling your name differently.\n      </p>\n    }\n    hintId=\"hintId5\"\n    errorId=\"errorId5\"\n    error=\"Select 'yes' if you have changed your name.\"\n  >\n    <Radios big>\n      <RadioBlock\n        label=\"Yes\"\n        id=\"anyRadioId9234\"\n        value=\"true\"\n        name=\"nameChange5\"\n        labelId=\"labelId9\"\n      />\n      <RadioBlock\n        label=\"No\"\n        id=\"anyRadioId10\"\n        value=\"false\"\n        name=\"nameChange5\"\n        labelId=\"labelId10\"\n      />\n    </Radios>\n  </FieldsetBlock>\n);\n",
-    "react-ts-styled-components": "import React from \"react\";\nimport FieldsetBlock from \"@govtnz/ds/build/react-ts-styled-components/FieldsetBlock\";\nimport H1 from \"@govtnz/ds/build/react-ts-styled-components/H1\";\nimport Radios from \"@govtnz/ds/build/react-ts-styled-components/Radios\";\nimport RadioBlock from \"@govtnz/ds/build/react-ts-styled-components/RadioBlock\";\n\nexport default () => (\n  <FieldsetBlock\n    legend={\n      <H1 styleSize=\"large\" id=\"nameChangeId5\">\n        Have you changed your name?\n      </H1>\n    }\n    hint={\n      <p>\n        This includes changing your last name or spelling your name differently.\n      </p>\n    }\n    hintId=\"hintId5\"\n    errorId=\"errorId5\"\n    error=\"Select 'yes' if you have changed your name.\"\n  >\n    <Radios big>\n      <RadioBlock\n        label=\"Yes\"\n        id=\"anyRadioId9234\"\n        value=\"true\"\n        name=\"nameChange5\"\n        labelId=\"labelId9\"\n      />\n      <RadioBlock\n        label=\"No\"\n        id=\"anyRadioId10\"\n        value=\"false\"\n        name=\"nameChange5\"\n        labelId=\"labelId10\"\n      />\n    </Radios>\n  </FieldsetBlock>\n);\n",
-    "vue-js": "<template>\n  <fieldset-block\n    legend=\"Have you changed your name?\"\n    hint=\"This includes changing your last name or spelling your name differently.\"\n    hintId=\"hintId5\"\n    errorId=\"errorId5\"\n    error=\"Select 'yes' if you have changed your name.\"\n  >\n    <c-radios big>\n      <radio-block\n        label=\"Yes\"\n        id=\"anyRadioId9234\"\n        value=\"true\"\n        name=\"nameChange5\"\n        labelId=\"labelId9\"\n      />\n      <radio-block\n        label=\"No\"\n        id=\"anyRadioId10\"\n        value=\"false\"\n        name=\"nameChange5\"\n        labelId=\"labelId10\"\n      />\n    </c-radios>\n  </fieldset-block>\n</template>\n<script>\nimport Vue from \"vue\";\nimport FieldsetBlock from \"@govtnz/ds/build/vue-js/FieldsetBlock.vue\";\nimport H1 from \"@govtnz/ds/build/vue-js/H1.vue\";\nimport Radios from \"@govtnz/ds/build/vue-js/Radios.vue\";\nimport RadioBlock from \"@govtnz/ds/build/vue-js/RadioBlock.vue\";\n\nexport default {\n  components: {\n    \"fieldset-block\": FieldsetBlock,\n    \"c-h1\": H1,\n    \"c-radios\": Radios,\n    \"radio-block\": RadioBlock\n  }\n};\n</script>\n",
-    "vue-ts": "<template>\n  <fieldset-block\n    legend=\"Have you changed your name?\"\n    hint=\"This includes changing your last name or spelling your name differently.\"\n    hintId=\"hintId5\"\n    errorId=\"errorId5\"\n    error=\"Select 'yes' if you have changed your name.\"\n  >\n    <c-radios big>\n      <radio-block\n        label=\"Yes\"\n        id=\"anyRadioId9234\"\n        value=\"true\"\n        name=\"nameChange5\"\n        labelId=\"labelId9\"\n      />\n      <radio-block\n        label=\"No\"\n        id=\"anyRadioId10\"\n        value=\"false\"\n        name=\"nameChange5\"\n        labelId=\"labelId10\"\n      />\n    </c-radios>\n  </fieldset-block>\n</template>\n<script lang=\"ts\">\nimport Vue from \"vue\";\nimport FieldsetBlock from \"@govtnz/ds/build/vue-ts/FieldsetBlock.vue\";\nimport H1 from \"@govtnz/ds/build/vue-ts/H1.vue\";\nimport Radios from \"@govtnz/ds/build/vue-ts/Radios.vue\";\nimport RadioBlock from \"@govtnz/ds/build/vue-ts/RadioBlock.vue\";\n\nexport default {\n  components: {\n    \"fieldset-block\": FieldsetBlock,\n    \"c-h1\": H1,\n    \"c-radios\": Radios,\n    \"radio-block\": RadioBlock\n  }\n};\n</script>\n",
-    "scss": "",
-    "html": "<!--\nRemember to add these styles:\nin CSS: FieldsetBlock.css, H1.css, Radios.css, RadioBlock.css\nOR in Sass (SCSS): FieldsetBlock.scss, H1.scss, Radios.scss, RadioBlock.scss\n-->\n<div class=\"g-fieldsetBlock-form-group\">\n  <fieldset\n    class=\"g-fieldsetBlock-fieldset\"\n    aria-describedby=\"hintId5 errorId5\"\n  >\n    <legend class=\"g-fieldsetBlock-fieldset__legend\">\n      <h1 class=\"g-heading g-heading-l\" id=\"nameChangeId5\">\n        Have you changed your name?\n      </h1>\n    </legend>\n    <div class=\"g-fieldsetBlock-hint\" id=\"hintId5\">\n      <p>\n        This includes changing your last name or spelling your name differently.\n      </p>\n    </div>\n    <div class=\"g-fieldsetBlock-error-message\" id=\"errorId5\">\n      <span class=\"g-fieldsetBlock-visually-hidden\"> Error: </span> Select 'yes'\n      if you have changed your name.\n    </div>\n    <div>\n      <div class=\"g-radios g-radios--big\">\n        <div class=\"g-radios__item\">\n          <input\n            class=\"g-radios__input\"\n            id=\"anyRadioId9234\"\n            name=\"nameChange5\"\n            type=\"radio\"\n            value=\"true\"\n          />\n          <label class=\"g-radios__label\" for=\"anyRadioId9234\"> Yes </label>\n        </div>\n        <div class=\"g-radios__item\">\n          <input\n            class=\"g-radios__input\"\n            id=\"anyRadioId10\"\n            name=\"nameChange5\"\n            type=\"radio\"\n            value=\"false\"\n          />\n          <label class=\"g-radios__label\" for=\"anyRadioId10\"> No </label>\n        </div>\n      </div>\n    </div>\n  </fieldset>\n</div>\n",
-    "css": "",
-    "twig-embed": "{% embed \"FieldsetBlock.html.twig\"  with {'hintId':'hintId5', 'errorId':'errorId5', 'error':'Select 'yes' if you have changed your name.'} only %}{% block legend %}{% embed \"H1.html.twig\"  with {'styleSize':'large', 'id':'nameChangeId5', 'children':'Have you changed your name?'} only %}{% endembed %}{% endblock %}{% block hint %}<p >This includes changing your last name or spelling your name differently.</p>{% endblock %}{% block children %}{% embed \"Radios.html.twig\"  with {'big':'true'} only %}{% block children %}{% embed \"RadioBlock.html.twig\"  with {'label':'Yes', 'id':'anyRadioId9234', 'value':'true', 'name':'nameChange5', 'labelId':'labelId9'} only %}{% endembed %}\n{% embed \"RadioBlock.html.twig\"  with {'label':'No', 'id':'anyRadioId10', 'value':'false', 'name':'nameChange5', 'labelId':'labelId10'} only %}{% endembed %}{% endblock %}{% endembed %}{% endblock %}{% endembed %}",
-    "angular": ""
-  },
-  {
-    "mustache": "// Developer note: ensure your \".mustache\" files are imported as plain text. In Webpack you might use https://github.com/webpack-contrib/raw-loader\nimport Mustache from \"mustache\";\nimport FieldsetBlock from \"@govtnz/ds/build/mustache/FieldsetBlock.mustache\";\nimport H1 from \"@govtnz/ds/build/mustache/H1.mustache\";\nimport Radios from \"@govtnz/ds/build/mustache/Radios.mustache\";\nimport RadioBlock from \"@govtnz/ds/build/mustache/RadioBlock.mustache\";\n\n// The variables FieldsetBlock, H1, Radios, RadioBlock are all strings that are mustache templates.\n\nexport default `${Mustache.render(FieldsetBlock, {\n  legend: `${Mustache.render(H1, {\n    \"styleSize=large\": true,\n    styleSize: true,\n    id: `nameChangeId1`,\n    children: `Have you changed your name?`\n  })}`,\n  hint: `<p >This includes changing your last name or spelling your name differently.</p>`,\n  hintId: `hintId1`,\n  children: `${Mustache.render(Radios, {\n    big: ``,\n    children: `${Mustache.render(RadioBlock, {\n      label: `Yes`,\n      id: `anyRadioId`,\n      value: `true`,\n      name: `nameChange1`,\n      labelId: `labelId1`\n    })}${Mustache.render(RadioBlock, {\n      label: `No`,\n      id: `anyRadioId255`,\n      value: `false`,\n      name: `nameChange1`,\n      labelId: `labelId255`\n    })}`\n  })}`\n})}`;\n",
-    "silverstripe-components": "",
-    "react-js": "import React from \"react\";\nimport FieldsetBlock from \"@govtnz/ds/build/react-js/FieldsetBlock\";\nimport \"@govtnz/ds/build/css/FieldsetBlock.css\"; // or @govtnz/ds/build/scss/FieldsetBlock.scss\nimport H1 from \"@govtnz/ds/build/react-js/H1\";\nimport \"@govtnz/ds/build/css/H1.css\"; // or @govtnz/ds/build/scss/H1.scss\nimport Radios from \"@govtnz/ds/build/react-js/Radios\";\nimport \"@govtnz/ds/build/css/Radios.css\"; // or @govtnz/ds/build/scss/Radios.scss\nimport RadioBlock from \"@govtnz/ds/build/react-js/RadioBlock\";\nimport \"@govtnz/ds/build/css/RadioBlock.css\"; // or @govtnz/ds/build/scss/RadioBlock.scss\n\nexport default () => (\n  <FieldsetBlock\n    legend={\n      <H1 styleSize=\"large\" id=\"nameChangeId1\">\n        Have you changed your name?\n      </H1>\n    }\n    hint={\n      <p>\n        This includes changing your last name or spelling your name differently.\n      </p>\n    }\n    hintId=\"hintId1\"\n  >\n    <Radios big>\n      <RadioBlock\n        label=\"Yes\"\n        id=\"anyRadioId\"\n        value=\"true\"\n        name=\"nameChange1\"\n        labelId=\"labelId1\"\n      />\n      <RadioBlock\n        label=\"No\"\n        id=\"anyRadioId255\"\n        value=\"false\"\n        name=\"nameChange1\"\n        labelId=\"labelId255\"\n      />\n    </Radios>\n  </FieldsetBlock>\n);\n",
-    "react-ts": "import React from \"react\";\nimport FieldsetBlock from \"@govtnz/ds/build/react-ts/FieldsetBlock\";\nimport \"@govtnz/ds/build/css/FieldsetBlock.css\"; // or @govtnz/ds/build/scss/FieldsetBlock.scss\nimport H1 from \"@govtnz/ds/build/react-ts/H1\";\nimport \"@govtnz/ds/build/css/H1.css\"; // or @govtnz/ds/build/scss/H1.scss\nimport Radios from \"@govtnz/ds/build/react-ts/Radios\";\nimport \"@govtnz/ds/build/css/Radios.css\"; // or @govtnz/ds/build/scss/Radios.scss\nimport RadioBlock from \"@govtnz/ds/build/react-ts/RadioBlock\";\nimport \"@govtnz/ds/build/css/RadioBlock.css\"; // or @govtnz/ds/build/scss/RadioBlock.scss\n\nexport default () => (\n  <FieldsetBlock\n    legend={\n      <H1 styleSize=\"large\" id=\"nameChangeId1\">\n        Have you changed your name?\n      </H1>\n    }\n    hint={\n      <p>\n        This includes changing your last name or spelling your name differently.\n      </p>\n    }\n    hintId=\"hintId1\"\n  >\n    <Radios big>\n      <RadioBlock\n        label=\"Yes\"\n        id=\"anyRadioId\"\n        value=\"true\"\n        name=\"nameChange1\"\n        labelId=\"labelId1\"\n      />\n      <RadioBlock\n        label=\"No\"\n        id=\"anyRadioId255\"\n        value=\"false\"\n        name=\"nameChange1\"\n        labelId=\"labelId255\"\n      />\n    </Radios>\n  </FieldsetBlock>\n);\n",
-    "react-js-styled-components": "import React from \"react\";\nimport FieldsetBlock from \"@govtnz/ds/build/react-js-styled-components/FieldsetBlock\";\nimport H1 from \"@govtnz/ds/build/react-js-styled-components/H1\";\nimport Radios from \"@govtnz/ds/build/react-js-styled-components/Radios\";\nimport RadioBlock from \"@govtnz/ds/build/react-js-styled-components/RadioBlock\";\n\nexport default () => (\n  <FieldsetBlock\n    legend={\n      <H1 styleSize=\"large\" id=\"nameChangeId1\">\n        Have you changed your name?\n      </H1>\n    }\n    hint={\n      <p>\n        This includes changing your last name or spelling your name differently.\n      </p>\n    }\n    hintId=\"hintId1\"\n  >\n    <Radios big>\n      <RadioBlock\n        label=\"Yes\"\n        id=\"anyRadioId\"\n        value=\"true\"\n        name=\"nameChange1\"\n        labelId=\"labelId1\"\n      />\n      <RadioBlock\n        label=\"No\"\n        id=\"anyRadioId255\"\n        value=\"false\"\n        name=\"nameChange1\"\n        labelId=\"labelId255\"\n      />\n    </Radios>\n  </FieldsetBlock>\n);\n",
-    "react-ts-styled-components": "import React from \"react\";\nimport FieldsetBlock from \"@govtnz/ds/build/react-ts-styled-components/FieldsetBlock\";\nimport H1 from \"@govtnz/ds/build/react-ts-styled-components/H1\";\nimport Radios from \"@govtnz/ds/build/react-ts-styled-components/Radios\";\nimport RadioBlock from \"@govtnz/ds/build/react-ts-styled-components/RadioBlock\";\n\nexport default () => (\n  <FieldsetBlock\n    legend={\n      <H1 styleSize=\"large\" id=\"nameChangeId1\">\n        Have you changed your name?\n      </H1>\n    }\n    hint={\n      <p>\n        This includes changing your last name or spelling your name differently.\n      </p>\n    }\n    hintId=\"hintId1\"\n  >\n    <Radios big>\n      <RadioBlock\n        label=\"Yes\"\n        id=\"anyRadioId\"\n        value=\"true\"\n        name=\"nameChange1\"\n        labelId=\"labelId1\"\n      />\n      <RadioBlock\n        label=\"No\"\n        id=\"anyRadioId255\"\n        value=\"false\"\n        name=\"nameChange1\"\n        labelId=\"labelId255\"\n      />\n    </Radios>\n  </FieldsetBlock>\n);\n",
-    "vue-js": "<template>\n  <fieldset-block\n    legend=\"Have you changed your name?\"\n    hint=\"This includes changing your last name or spelling your name differently.\"\n    hintId=\"hintId1\"\n  >\n    <c-radios big>\n      <radio-block\n        label=\"Yes\"\n        id=\"anyRadioId\"\n        value=\"true\"\n        name=\"nameChange1\"\n        labelId=\"labelId1\"\n      />\n      <radio-block\n        label=\"No\"\n        id=\"anyRadioId255\"\n        value=\"false\"\n        name=\"nameChange1\"\n        labelId=\"labelId255\"\n      />\n    </c-radios>\n  </fieldset-block>\n</template>\n<script>\nimport Vue from \"vue\";\nimport FieldsetBlock from \"@govtnz/ds/build/vue-js/FieldsetBlock.vue\";\nimport H1 from \"@govtnz/ds/build/vue-js/H1.vue\";\nimport Radios from \"@govtnz/ds/build/vue-js/Radios.vue\";\nimport RadioBlock from \"@govtnz/ds/build/vue-js/RadioBlock.vue\";\n\nexport default {\n  components: {\n    \"fieldset-block\": FieldsetBlock,\n    \"c-h1\": H1,\n    \"c-radios\": Radios,\n    \"radio-block\": RadioBlock\n  }\n};\n</script>\n",
-    "vue-ts": "<template>\n  <fieldset-block\n    legend=\"Have you changed your name?\"\n    hint=\"This includes changing your last name or spelling your name differently.\"\n    hintId=\"hintId1\"\n  >\n    <c-radios big>\n      <radio-block\n        label=\"Yes\"\n        id=\"anyRadioId\"\n        value=\"true\"\n        name=\"nameChange1\"\n        labelId=\"labelId1\"\n      />\n      <radio-block\n        label=\"No\"\n        id=\"anyRadioId255\"\n        value=\"false\"\n        name=\"nameChange1\"\n        labelId=\"labelId255\"\n      />\n    </c-radios>\n  </fieldset-block>\n</template>\n<script lang=\"ts\">\nimport Vue from \"vue\";\nimport FieldsetBlock from \"@govtnz/ds/build/vue-ts/FieldsetBlock.vue\";\nimport H1 from \"@govtnz/ds/build/vue-ts/H1.vue\";\nimport Radios from \"@govtnz/ds/build/vue-ts/Radios.vue\";\nimport RadioBlock from \"@govtnz/ds/build/vue-ts/RadioBlock.vue\";\n\nexport default {\n  components: {\n    \"fieldset-block\": FieldsetBlock,\n    \"c-h1\": H1,\n    \"c-radios\": Radios,\n    \"radio-block\": RadioBlock\n  }\n};\n</script>\n",
-    "scss": "",
-    "html": "<!--\nRemember to add these styles:\nin CSS: FieldsetBlock.css, H1.css, Radios.css, RadioBlock.css\nOR in Sass (SCSS): FieldsetBlock.scss, H1.scss, Radios.scss, RadioBlock.scss\n-->\n<div class=\"g-fieldsetBlock-form-group\">\n  <fieldset class=\"g-fieldsetBlock-fieldset\" aria-describedby=\"hintId1\">\n    <legend class=\"g-fieldsetBlock-fieldset__legend\">\n      <h1 class=\"g-heading g-heading-l\" id=\"nameChangeId1\">\n        Have you changed your name?\n      </h1>\n    </legend>\n    <div class=\"g-fieldsetBlock-hint\" id=\"hintId1\">\n      <p>\n        This includes changing your last name or spelling your name differently.\n      </p>\n    </div>\n    <div>\n      <div class=\"g-radios g-radios--big\">\n        <div class=\"g-radios__item\">\n          <input\n            class=\"g-radios__input\"\n            id=\"anyRadioId\"\n            name=\"nameChange1\"\n            type=\"radio\"\n            value=\"true\"\n          />\n          <label class=\"g-radios__label\" for=\"anyRadioId\"> Yes </label>\n        </div>\n        <div class=\"g-radios__item\">\n          <input\n            class=\"g-radios__input\"\n            id=\"anyRadioId255\"\n            name=\"nameChange1\"\n            type=\"radio\"\n            value=\"false\"\n          />\n          <label class=\"g-radios__label\" for=\"anyRadioId255\"> No </label>\n        </div>\n      </div>\n    </div>\n  </fieldset>\n</div>\n",
-    "css": "",
-    "twig-embed": "{% embed \"FieldsetBlock.html.twig\"  with {'hintId':'hintId1'} only %}{% block legend %}{% embed \"H1.html.twig\"  with {'styleSize':'large', 'id':'nameChangeId1', 'children':'Have you changed your name?'} only %}{% endembed %}{% endblock %}{% block hint %}<p >This includes changing your last name or spelling your name differently.</p>{% endblock %}{% block children %}{% embed \"Radios.html.twig\"  with {'big':'true'} only %}{% block children %}{% embed \"RadioBlock.html.twig\"  with {'label':'Yes', 'id':'anyRadioId', 'value':'true', 'name':'nameChange1', 'labelId':'labelId1'} only %}{% endembed %}\n{% embed \"RadioBlock.html.twig\"  with {'label':'No', 'id':'anyRadioId255', 'value':'false', 'name':'nameChange1', 'labelId':'labelId255'} only %}{% endembed %}{% endblock %}{% endembed %}{% endblock %}{% endembed %}",
-    "angular": ""
-  },
-  {
-    "mustache": "// Developer note: ensure your \".mustache\" files are imported as plain text. In Webpack you might use https://github.com/webpack-contrib/raw-loader\nimport Mustache from \"mustache\";\nimport FieldsetBlock from \"@govtnz/ds/build/mustache/FieldsetBlock.mustache\";\nimport H1 from \"@govtnz/ds/build/mustache/H1.mustache\";\nimport Radios from \"@govtnz/ds/build/mustache/Radios.mustache\";\nimport RadioBlock from \"@govtnz/ds/build/mustache/RadioBlock.mustache\";\n\n// The variables FieldsetBlock, H1, Radios, RadioBlock are all strings that are mustache templates.\n\nexport default `${Mustache.render(FieldsetBlock, {\n  legend: `${Mustache.render(H1, {\n    \"styleSize=large\": true,\n    styleSize: true,\n    id: `nameChangeId5`,\n    children: `Have you changed your name?`\n  })}`,\n  hint: `<p >This includes changing your last name or spelling your name differently.</p>`,\n  hintId: `hintId5`,\n  errorId: true,\n  error: `Select 'yes' if you have changed your name.`,\n  children: `${Mustache.render(Radios, {\n    inline: true,\n    big: ``,\n    children: `${Mustache.render(RadioBlock, {\n      label: `Yes`,\n      id: `anyRadioId9234`,\n      value: `true`,\n      name: `nameChange5`,\n      labelId: `labelId9`\n    })}${Mustache.render(RadioBlock, {\n      label: `No`,\n      id: `anyRadioId10`,\n      value: `false`,\n      name: `nameChange5`,\n      labelId: `labelId10`\n    })}`\n  })}`\n})}`;\n",
-    "silverstripe-components": "",
-    "react-js": "import React from \"react\";\nimport FieldsetBlock from \"@govtnz/ds/build/react-js/FieldsetBlock\";\nimport \"@govtnz/ds/build/css/FieldsetBlock.css\"; // or @govtnz/ds/build/scss/FieldsetBlock.scss\nimport H1 from \"@govtnz/ds/build/react-js/H1\";\nimport \"@govtnz/ds/build/css/H1.css\"; // or @govtnz/ds/build/scss/H1.scss\nimport Radios from \"@govtnz/ds/build/react-js/Radios\";\nimport \"@govtnz/ds/build/css/Radios.css\"; // or @govtnz/ds/build/scss/Radios.scss\nimport RadioBlock from \"@govtnz/ds/build/react-js/RadioBlock\";\nimport \"@govtnz/ds/build/css/RadioBlock.css\"; // or @govtnz/ds/build/scss/RadioBlock.scss\n\nexport default () => (\n  <FieldsetBlock\n    legend={\n      <H1 styleSize=\"large\" id=\"nameChangeId5\">\n        Have you changed your name?\n      </H1>\n    }\n    hint={\n      <p>\n        This includes changing your last name or spelling your name differently.\n      </p>\n    }\n    hintId=\"hintId5\"\n    errorId=\"errorId5\"\n    error=\"Select 'yes' if you have changed your name.\"\n  >\n    <Radios inline big>\n      <RadioBlock\n        label=\"Yes\"\n        id=\"anyRadioId9234\"\n        value=\"true\"\n        name=\"nameChange5\"\n        labelId=\"labelId9\"\n      />\n      <RadioBlock\n        label=\"No\"\n        id=\"anyRadioId10\"\n        value=\"false\"\n        name=\"nameChange5\"\n        labelId=\"labelId10\"\n      />\n    </Radios>\n  </FieldsetBlock>\n);\n",
-    "react-ts": "import React from \"react\";\nimport FieldsetBlock from \"@govtnz/ds/build/react-ts/FieldsetBlock\";\nimport \"@govtnz/ds/build/css/FieldsetBlock.css\"; // or @govtnz/ds/build/scss/FieldsetBlock.scss\nimport H1 from \"@govtnz/ds/build/react-ts/H1\";\nimport \"@govtnz/ds/build/css/H1.css\"; // or @govtnz/ds/build/scss/H1.scss\nimport Radios from \"@govtnz/ds/build/react-ts/Radios\";\nimport \"@govtnz/ds/build/css/Radios.css\"; // or @govtnz/ds/build/scss/Radios.scss\nimport RadioBlock from \"@govtnz/ds/build/react-ts/RadioBlock\";\nimport \"@govtnz/ds/build/css/RadioBlock.css\"; // or @govtnz/ds/build/scss/RadioBlock.scss\n\nexport default () => (\n  <FieldsetBlock\n    legend={\n      <H1 styleSize=\"large\" id=\"nameChangeId5\">\n        Have you changed your name?\n      </H1>\n    }\n    hint={\n      <p>\n        This includes changing your last name or spelling your name differently.\n      </p>\n    }\n    hintId=\"hintId5\"\n    errorId=\"errorId5\"\n    error=\"Select 'yes' if you have changed your name.\"\n  >\n    <Radios inline big>\n      <RadioBlock\n        label=\"Yes\"\n        id=\"anyRadioId9234\"\n        value=\"true\"\n        name=\"nameChange5\"\n        labelId=\"labelId9\"\n      />\n      <RadioBlock\n        label=\"No\"\n        id=\"anyRadioId10\"\n        value=\"false\"\n        name=\"nameChange5\"\n        labelId=\"labelId10\"\n      />\n    </Radios>\n  </FieldsetBlock>\n);\n",
-    "react-js-styled-components": "import React from \"react\";\nimport FieldsetBlock from \"@govtnz/ds/build/react-js-styled-components/FieldsetBlock\";\nimport H1 from \"@govtnz/ds/build/react-js-styled-components/H1\";\nimport Radios from \"@govtnz/ds/build/react-js-styled-components/Radios\";\nimport RadioBlock from \"@govtnz/ds/build/react-js-styled-components/RadioBlock\";\n\nexport default () => (\n  <FieldsetBlock\n    legend={\n      <H1 styleSize=\"large\" id=\"nameChangeId5\">\n        Have you changed your name?\n      </H1>\n    }\n    hint={\n      <p>\n        This includes changing your last name or spelling your name differently.\n      </p>\n    }\n    hintId=\"hintId5\"\n    errorId=\"errorId5\"\n    error=\"Select 'yes' if you have changed your name.\"\n  >\n    <Radios inline big>\n      <RadioBlock\n        label=\"Yes\"\n        id=\"anyRadioId9234\"\n        value=\"true\"\n        name=\"nameChange5\"\n        labelId=\"labelId9\"\n      />\n      <RadioBlock\n        label=\"No\"\n        id=\"anyRadioId10\"\n        value=\"false\"\n        name=\"nameChange5\"\n        labelId=\"labelId10\"\n      />\n    </Radios>\n  </FieldsetBlock>\n);\n",
-    "react-ts-styled-components": "import React from \"react\";\nimport FieldsetBlock from \"@govtnz/ds/build/react-ts-styled-components/FieldsetBlock\";\nimport H1 from \"@govtnz/ds/build/react-ts-styled-components/H1\";\nimport Radios from \"@govtnz/ds/build/react-ts-styled-components/Radios\";\nimport RadioBlock from \"@govtnz/ds/build/react-ts-styled-components/RadioBlock\";\n\nexport default () => (\n  <FieldsetBlock\n    legend={\n      <H1 styleSize=\"large\" id=\"nameChangeId5\">\n        Have you changed your name?\n      </H1>\n    }\n    hint={\n      <p>\n        This includes changing your last name or spelling your name differently.\n      </p>\n    }\n    hintId=\"hintId5\"\n    errorId=\"errorId5\"\n    error=\"Select 'yes' if you have changed your name.\"\n  >\n    <Radios inline big>\n      <RadioBlock\n        label=\"Yes\"\n        id=\"anyRadioId9234\"\n        value=\"true\"\n        name=\"nameChange5\"\n        labelId=\"labelId9\"\n      />\n      <RadioBlock\n        label=\"No\"\n        id=\"anyRadioId10\"\n        value=\"false\"\n        name=\"nameChange5\"\n        labelId=\"labelId10\"\n      />\n    </Radios>\n  </FieldsetBlock>\n);\n",
-    "vue-js": "<template>\n  <fieldset-block\n    legend=\"Have you changed your name?\"\n    hint=\"This includes changing your last name or spelling your name differently.\"\n    hintId=\"hintId5\"\n    errorId=\"errorId5\"\n    error=\"Select 'yes' if you have changed your name.\"\n  >\n    <c-radios inline big>\n      <radio-block\n        label=\"Yes\"\n        id=\"anyRadioId9234\"\n        value=\"true\"\n        name=\"nameChange5\"\n        labelId=\"labelId9\"\n      />\n      <radio-block\n        label=\"No\"\n        id=\"anyRadioId10\"\n        value=\"false\"\n        name=\"nameChange5\"\n        labelId=\"labelId10\"\n      />\n    </c-radios>\n  </fieldset-block>\n</template>\n<script>\nimport Vue from \"vue\";\nimport FieldsetBlock from \"@govtnz/ds/build/vue-js/FieldsetBlock.vue\";\nimport H1 from \"@govtnz/ds/build/vue-js/H1.vue\";\nimport Radios from \"@govtnz/ds/build/vue-js/Radios.vue\";\nimport RadioBlock from \"@govtnz/ds/build/vue-js/RadioBlock.vue\";\n\nexport default {\n  components: {\n    \"fieldset-block\": FieldsetBlock,\n    \"c-h1\": H1,\n    \"c-radios\": Radios,\n    \"radio-block\": RadioBlock\n  }\n};\n</script>\n",
-    "vue-ts": "<template>\n  <fieldset-block\n    legend=\"Have you changed your name?\"\n    hint=\"This includes changing your last name or spelling your name differently.\"\n    hintId=\"hintId5\"\n    errorId=\"errorId5\"\n    error=\"Select 'yes' if you have changed your name.\"\n  >\n    <c-radios inline big>\n      <radio-block\n        label=\"Yes\"\n        id=\"anyRadioId9234\"\n        value=\"true\"\n        name=\"nameChange5\"\n        labelId=\"labelId9\"\n      />\n      <radio-block\n        label=\"No\"\n        id=\"anyRadioId10\"\n        value=\"false\"\n        name=\"nameChange5\"\n        labelId=\"labelId10\"\n      />\n    </c-radios>\n  </fieldset-block>\n</template>\n<script lang=\"ts\">\nimport Vue from \"vue\";\nimport FieldsetBlock from \"@govtnz/ds/build/vue-ts/FieldsetBlock.vue\";\nimport H1 from \"@govtnz/ds/build/vue-ts/H1.vue\";\nimport Radios from \"@govtnz/ds/build/vue-ts/Radios.vue\";\nimport RadioBlock from \"@govtnz/ds/build/vue-ts/RadioBlock.vue\";\n\nexport default {\n  components: {\n    \"fieldset-block\": FieldsetBlock,\n    \"c-h1\": H1,\n    \"c-radios\": Radios,\n    \"radio-block\": RadioBlock\n  }\n};\n</script>\n",
-    "scss": "",
-    "html": "<!--\nRemember to add these styles:\nin CSS: FieldsetBlock.css, H1.css, Radios.css, RadioBlock.css\nOR in Sass (SCSS): FieldsetBlock.scss, H1.scss, Radios.scss, RadioBlock.scss\n-->\n<div class=\"g-fieldsetBlock-form-group\">\n  <fieldset\n    class=\"g-fieldsetBlock-fieldset\"\n    aria-describedby=\"hintId5 errorId5\"\n  >\n    <legend class=\"g-fieldsetBlock-fieldset__legend\">\n      <h1 class=\"g-heading g-heading-l\" id=\"nameChangeId5\">\n        Have you changed your name?\n      </h1>\n    </legend>\n    <div class=\"g-fieldsetBlock-hint\" id=\"hintId5\">\n      <p>\n        This includes changing your last name or spelling your name differently.\n      </p>\n    </div>\n    <div class=\"g-fieldsetBlock-error-message\" id=\"errorId5\">\n      <span class=\"g-fieldsetBlock-visually-hidden\"> Error: </span> Select 'yes'\n      if you have changed your name.\n    </div>\n    <div>\n      <div class=\"g-radios g-radios--inline g-radios--big\">\n        <div class=\"g-radios__item\">\n          <input\n            class=\"g-radios__input\"\n            id=\"anyRadioId9234\"\n            name=\"nameChange5\"\n            type=\"radio\"\n            value=\"true\"\n          />\n          <label class=\"g-radios__label\" for=\"anyRadioId9234\"> Yes </label>\n        </div>\n        <div class=\"g-radios__item\">\n          <input\n            class=\"g-radios__input\"\n            id=\"anyRadioId10\"\n            name=\"nameChange5\"\n            type=\"radio\"\n            value=\"false\"\n          />\n          <label class=\"g-radios__label\" for=\"anyRadioId10\"> No </label>\n        </div>\n      </div>\n    </div>\n  </fieldset>\n</div>\n",
-    "css": "",
-    "twig-embed": "{% embed \"FieldsetBlock.html.twig\"  with {'hintId':'hintId5', 'errorId':'errorId5', 'error':'Select 'yes' if you have changed your name.'} only %}{% block legend %}{% embed \"H1.html.twig\"  with {'styleSize':'large', 'id':'nameChangeId5', 'children':'Have you changed your name?'} only %}{% endembed %}{% endblock %}{% block hint %}<p >This includes changing your last name or spelling your name differently.</p>{% endblock %}{% block children %}{% embed \"Radios.html.twig\"  with {'inline':'true', 'big':'true'} only %}{% block children %}{% embed \"RadioBlock.html.twig\"  with {'label':'Yes', 'id':'anyRadioId9234', 'value':'true', 'name':'nameChange5', 'labelId':'labelId9'} only %}{% endembed %}\n{% embed \"RadioBlock.html.twig\"  with {'label':'No', 'id':'anyRadioId10', 'value':'false', 'name':'nameChange5', 'labelId':'labelId10'} only %}{% endembed %}{% endblock %}{% endembed %}{% endblock %}{% endembed %}",
-    "angular": ""
-  },
-  {
     "mustache": "// Developer note: ensure your \".mustache\" files are imported as plain text. In Webpack you might use https://github.com/webpack-contrib/raw-loader\nimport Mustache from \"mustache\";\nimport FieldsetBlock from \"@govtnz/ds/build/mustache/FieldsetBlock.mustache\";\nimport H1 from \"@govtnz/ds/build/mustache/H1.mustache\";\nimport RadioBlock from \"@govtnz/ds/build/mustache/RadioBlock.mustache\";\n\n// The variables FieldsetBlock, H1, RadioBlock are all strings that are mustache templates.\n\nexport default `${Mustache.render(FieldsetBlock, {\n  legend: `${Mustache.render(H1, {\n    \"styleSize=large\": true,\n    styleSize: true,\n    id: `whereLiveTitle`,\n    children: `Where do you live?`\n  })}`,\n  children: `${Mustache.render(RadioBlock, {\n    label: `North Island`,\n    id: `anyRadioId3334`,\n    name: `where`,\n    value: `north`,\n    labelId: `labelId3`\n  })}${Mustache.render(RadioBlock, {\n    label: `South Island`,\n    id: `anyRadioId42344`,\n    name: `where`,\n    value: `south`,\n    labelId: `labelId4`\n  })}${Mustache.render(RadioBlock, {\n    label: `Stewart Island`,\n    id: `anyRadioId5234`,\n    name: `where`,\n    value: `stewart`,\n    labelId: `labelId5`\n  })}${Mustache.render(RadioBlock, {\n    label: `Chatham Islands`,\n    id: `anyRadioId6234`,\n    name: `where`,\n    value: `chatham`,\n    labelId: `labelId6`\n  })}`\n})}`;\n",
     "silverstripe-components": "",
     "react-js": "import React from \"react\";\nimport FieldsetBlock from \"@govtnz/ds/build/react-js/FieldsetBlock\";\nimport \"@govtnz/ds/build/css/FieldsetBlock.css\"; // or @govtnz/ds/build/scss/FieldsetBlock.scss\nimport H1 from \"@govtnz/ds/build/react-js/H1\";\nimport \"@govtnz/ds/build/css/H1.css\"; // or @govtnz/ds/build/scss/H1.scss\nimport RadioBlock from \"@govtnz/ds/build/react-js/RadioBlock\";\nimport \"@govtnz/ds/build/css/RadioBlock.css\"; // or @govtnz/ds/build/scss/RadioBlock.scss\n\nexport default () => (\n  <FieldsetBlock\n    legend={\n      <H1 styleSize=\"large\" id=\"whereLiveTitle\">\n        Where do you live?\n      </H1>\n    }\n  >\n    <RadioBlock\n      label=\"North Island\"\n      id=\"anyRadioId3334\"\n      name=\"where\"\n      value=\"north\"\n      labelId=\"labelId3\"\n    />\n    <RadioBlock\n      label=\"South Island\"\n      id=\"anyRadioId42344\"\n      name=\"where\"\n      value=\"south\"\n      labelId=\"labelId4\"\n    />\n    <RadioBlock\n      label=\"Stewart Island\"\n      id=\"anyRadioId5234\"\n      name=\"where\"\n      value=\"stewart\"\n      labelId=\"labelId5\"\n    />\n    <RadioBlock\n      label=\"Chatham Islands\"\n      id=\"anyRadioId6234\"\n      name=\"where\"\n      value=\"chatham\"\n      labelId=\"labelId6\"\n    />\n  </FieldsetBlock>\n);\n",
@@ -150,6 +75,21 @@ export default [
     "html": "<!--\nRemember to add these styles:\nin CSS: FieldsetBlock.css, H1.css, RadioBlock.css, FormDivider.css\nOR in Sass (SCSS): FieldsetBlock.scss, H1.scss, RadioBlock.scss, FormDivider.scss\n-->\n<div class=\"g-fieldsetBlock-form-group\">\n  <fieldset class=\"g-fieldsetBlock-fieldset\" aria-describedby=\"hintId234\">\n    <legend class=\"g-fieldsetBlock-fieldset__legend\">\n      <h1 class=\"g-heading g-heading-l\" id=\"providerChoiceTitle2\">\n        Choose one provider\n      </h1>\n    </legend>\n    <div class=\"g-fieldsetBlock-hint\" id=\"hintId234\">\n      <p>Select one provider.</p>\n    </div>\n    <div>\n      <div class=\"g-radios__item\">\n        <input\n          class=\"g-radios__input\"\n          id=\"anyRadioId7345\"\n          name=\"providerChoice3\"\n          type=\"radio\"\n          aria-describedby=\"someHintId72344\"\n          value=\"provider1\"\n        />\n        <label class=\"g-radios__label\" for=\"anyRadioId7345\">\n          Sign in with Aardvark Services\n        </label>\n        <div class=\"g-radios__hint\" id=\"someHintId72344\">North Island only</div>\n      </div>\n      <div class=\"g-radios__item\">\n        <input\n          class=\"g-radios__input\"\n          id=\"anyRadioId8345\"\n          name=\"providerChoice3\"\n          type=\"radio\"\n          aria-describedby=\"someHintId8533\"\n          value=\"provider2\"\n        />\n        <label class=\"g-radios__label\" for=\"anyRadioId8345\">\n          Sign in with Dandelion Services\n        </label>\n        <div class=\"g-radios__hint\" id=\"someHintId8533\">\n          South and Stewart Islands only\n        </div>\n      </div>\n      <div class=\"g-form-divider\"><p>-or-</p></div>\n      <div class=\"g-radios__item\">\n        <input\n          class=\"g-radios__input\"\n          id=\"anyRadioId3459\"\n          name=\"providerChoice3\"\n          type=\"radio\"\n          aria-describedby=\"someHintId7234444\"\n          value=\"provider3\"\n        />\n        <label class=\"g-radios__label\" for=\"anyRadioId3459\">\n          Sign in with Eel Services\n        </label>\n        <div class=\"g-radios__hint\" id=\"someHintId7234444\">\n          Non-New Zealand residents only\n        </div>\n      </div>\n    </div>\n  </fieldset>\n</div>\n",
     "css": "",
     "twig-embed": "{% embed \"FieldsetBlock.html.twig\"  with {'hintId':'hintId234'} only %}{% block legend %}{% embed \"H1.html.twig\"  with {'styleSize':'large', 'id':'providerChoiceTitle2', 'children':'Choose one provider'} only %}{% endembed %}{% endblock %}{% block hint %}<p >Select one provider.</p>{% endblock %}{% block children %}{% embed \"RadioBlock.html.twig\"  with {'label':'Sign in with Aardvark Services', 'hint':'North Island only', 'hintId':'someHintId72344', 'id':'anyRadioId7345', 'name':'providerChoice3', 'value':'provider1', 'labelId':'labelId5'} only %}{% endembed %}\n{% embed \"RadioBlock.html.twig\"  with {'label':'Sign in with Dandelion Services', 'hint':'South and Stewart Islands only', 'hintId':'someHintId8533', 'id':'anyRadioId8345', 'name':'providerChoice3', 'value':'provider2', 'labelId':'labelId8'} only %}{% endembed %}\n{% embed \"FormDivider.html.twig\" %}{% block children %}<p >-or-</p>{% endblock %}{% endembed %}\n{% embed \"RadioBlock.html.twig\"  with {'label':'Sign in with Eel Services', 'hint':'Non-New Zealand residents only', 'hintId':'someHintId7234444', 'id':'anyRadioId3459', 'name':'providerChoice3', 'value':'provider3', 'labelId':'labelId9'} only %}{% endembed %}{% endblock %}{% endembed %}",
+    "angular": ""
+  },
+  {
+    "mustache": "// Developer note: ensure your \".mustache\" files are imported as plain text. In Webpack you might use https://github.com/webpack-contrib/raw-loader\nimport Mustache from \"mustache\";\nimport FieldsetBlock from \"@govtnz/ds/build/mustache/FieldsetBlock.mustache\";\nimport H1 from \"@govtnz/ds/build/mustache/H1.mustache\";\nimport Radios from \"@govtnz/ds/build/mustache/Radios.mustache\";\nimport RadioBlock from \"@govtnz/ds/build/mustache/RadioBlock.mustache\";\n\n// The variables FieldsetBlock, H1, Radios, RadioBlock are all strings that are mustache templates.\n\nexport default `${Mustache.render(FieldsetBlock, {\n  legend: `${Mustache.render(H1, {\n    \"styleSize=large\": true,\n    styleSize: true,\n    id: `nameChangeId1`,\n    children: `Have you changed your name?`\n  })}`,\n  hint: `<p >This includes changing your last name or spelling your name differently.</p>`,\n  hintId: `hintId1`,\n  children: `${Mustache.render(Radios, {\n    inline: true,\n    big: ``,\n    children: `${Mustache.render(RadioBlock, {\n      label: `Yes`,\n      id: `anyRadioId`,\n      value: `true`,\n      name: `nameChange1`,\n      labelId: `labelId1`\n    })}${Mustache.render(RadioBlock, {\n      label: `No`,\n      id: `anyRadioId255`,\n      value: `false`,\n      name: `nameChange1`,\n      labelId: `labelId255`\n    })}`\n  })}`\n})}`;\n",
+    "silverstripe-components": "",
+    "react-js": "import React from \"react\";\nimport FieldsetBlock from \"@govtnz/ds/build/react-js/FieldsetBlock\";\nimport \"@govtnz/ds/build/css/FieldsetBlock.css\"; // or @govtnz/ds/build/scss/FieldsetBlock.scss\nimport H1 from \"@govtnz/ds/build/react-js/H1\";\nimport \"@govtnz/ds/build/css/H1.css\"; // or @govtnz/ds/build/scss/H1.scss\nimport Radios from \"@govtnz/ds/build/react-js/Radios\";\nimport \"@govtnz/ds/build/css/Radios.css\"; // or @govtnz/ds/build/scss/Radios.scss\nimport RadioBlock from \"@govtnz/ds/build/react-js/RadioBlock\";\nimport \"@govtnz/ds/build/css/RadioBlock.css\"; // or @govtnz/ds/build/scss/RadioBlock.scss\n\nexport default () => (\n  <FieldsetBlock\n    legend={\n      <H1 styleSize=\"large\" id=\"nameChangeId1\">\n        Have you changed your name?\n      </H1>\n    }\n    hint={\n      <p>\n        This includes changing your last name or spelling your name differently.\n      </p>\n    }\n    hintId=\"hintId1\"\n  >\n    <Radios inline big>\n      <RadioBlock\n        label=\"Yes\"\n        id=\"anyRadioId\"\n        value=\"true\"\n        name=\"nameChange1\"\n        labelId=\"labelId1\"\n      />\n      <RadioBlock\n        label=\"No\"\n        id=\"anyRadioId255\"\n        value=\"false\"\n        name=\"nameChange1\"\n        labelId=\"labelId255\"\n      />\n    </Radios>\n  </FieldsetBlock>\n);\n",
+    "react-ts": "import React from \"react\";\nimport FieldsetBlock from \"@govtnz/ds/build/react-ts/FieldsetBlock\";\nimport \"@govtnz/ds/build/css/FieldsetBlock.css\"; // or @govtnz/ds/build/scss/FieldsetBlock.scss\nimport H1 from \"@govtnz/ds/build/react-ts/H1\";\nimport \"@govtnz/ds/build/css/H1.css\"; // or @govtnz/ds/build/scss/H1.scss\nimport Radios from \"@govtnz/ds/build/react-ts/Radios\";\nimport \"@govtnz/ds/build/css/Radios.css\"; // or @govtnz/ds/build/scss/Radios.scss\nimport RadioBlock from \"@govtnz/ds/build/react-ts/RadioBlock\";\nimport \"@govtnz/ds/build/css/RadioBlock.css\"; // or @govtnz/ds/build/scss/RadioBlock.scss\n\nexport default () => (\n  <FieldsetBlock\n    legend={\n      <H1 styleSize=\"large\" id=\"nameChangeId1\">\n        Have you changed your name?\n      </H1>\n    }\n    hint={\n      <p>\n        This includes changing your last name or spelling your name differently.\n      </p>\n    }\n    hintId=\"hintId1\"\n  >\n    <Radios inline big>\n      <RadioBlock\n        label=\"Yes\"\n        id=\"anyRadioId\"\n        value=\"true\"\n        name=\"nameChange1\"\n        labelId=\"labelId1\"\n      />\n      <RadioBlock\n        label=\"No\"\n        id=\"anyRadioId255\"\n        value=\"false\"\n        name=\"nameChange1\"\n        labelId=\"labelId255\"\n      />\n    </Radios>\n  </FieldsetBlock>\n);\n",
+    "react-js-styled-components": "import React from \"react\";\nimport FieldsetBlock from \"@govtnz/ds/build/react-js-styled-components/FieldsetBlock\";\nimport H1 from \"@govtnz/ds/build/react-js-styled-components/H1\";\nimport Radios from \"@govtnz/ds/build/react-js-styled-components/Radios\";\nimport RadioBlock from \"@govtnz/ds/build/react-js-styled-components/RadioBlock\";\n\nexport default () => (\n  <FieldsetBlock\n    legend={\n      <H1 styleSize=\"large\" id=\"nameChangeId1\">\n        Have you changed your name?\n      </H1>\n    }\n    hint={\n      <p>\n        This includes changing your last name or spelling your name differently.\n      </p>\n    }\n    hintId=\"hintId1\"\n  >\n    <Radios inline big>\n      <RadioBlock\n        label=\"Yes\"\n        id=\"anyRadioId\"\n        value=\"true\"\n        name=\"nameChange1\"\n        labelId=\"labelId1\"\n      />\n      <RadioBlock\n        label=\"No\"\n        id=\"anyRadioId255\"\n        value=\"false\"\n        name=\"nameChange1\"\n        labelId=\"labelId255\"\n      />\n    </Radios>\n  </FieldsetBlock>\n);\n",
+    "react-ts-styled-components": "import React from \"react\";\nimport FieldsetBlock from \"@govtnz/ds/build/react-ts-styled-components/FieldsetBlock\";\nimport H1 from \"@govtnz/ds/build/react-ts-styled-components/H1\";\nimport Radios from \"@govtnz/ds/build/react-ts-styled-components/Radios\";\nimport RadioBlock from \"@govtnz/ds/build/react-ts-styled-components/RadioBlock\";\n\nexport default () => (\n  <FieldsetBlock\n    legend={\n      <H1 styleSize=\"large\" id=\"nameChangeId1\">\n        Have you changed your name?\n      </H1>\n    }\n    hint={\n      <p>\n        This includes changing your last name or spelling your name differently.\n      </p>\n    }\n    hintId=\"hintId1\"\n  >\n    <Radios inline big>\n      <RadioBlock\n        label=\"Yes\"\n        id=\"anyRadioId\"\n        value=\"true\"\n        name=\"nameChange1\"\n        labelId=\"labelId1\"\n      />\n      <RadioBlock\n        label=\"No\"\n        id=\"anyRadioId255\"\n        value=\"false\"\n        name=\"nameChange1\"\n        labelId=\"labelId255\"\n      />\n    </Radios>\n  </FieldsetBlock>\n);\n",
+    "vue-js": "<template>\n  <fieldset-block\n    legend=\"Have you changed your name?\"\n    hint=\"This includes changing your last name or spelling your name differently.\"\n    hintId=\"hintId1\"\n  >\n    <c-radios inline big>\n      <radio-block\n        label=\"Yes\"\n        id=\"anyRadioId\"\n        value=\"true\"\n        name=\"nameChange1\"\n        labelId=\"labelId1\"\n      />\n      <radio-block\n        label=\"No\"\n        id=\"anyRadioId255\"\n        value=\"false\"\n        name=\"nameChange1\"\n        labelId=\"labelId255\"\n      />\n    </c-radios>\n  </fieldset-block>\n</template>\n<script>\nimport Vue from \"vue\";\nimport FieldsetBlock from \"@govtnz/ds/build/vue-js/FieldsetBlock.vue\";\nimport H1 from \"@govtnz/ds/build/vue-js/H1.vue\";\nimport Radios from \"@govtnz/ds/build/vue-js/Radios.vue\";\nimport RadioBlock from \"@govtnz/ds/build/vue-js/RadioBlock.vue\";\n\nexport default {\n  components: {\n    \"fieldset-block\": FieldsetBlock,\n    \"c-h1\": H1,\n    \"c-radios\": Radios,\n    \"radio-block\": RadioBlock\n  }\n};\n</script>\n",
+    "vue-ts": "<template>\n  <fieldset-block\n    legend=\"Have you changed your name?\"\n    hint=\"This includes changing your last name or spelling your name differently.\"\n    hintId=\"hintId1\"\n  >\n    <c-radios inline big>\n      <radio-block\n        label=\"Yes\"\n        id=\"anyRadioId\"\n        value=\"true\"\n        name=\"nameChange1\"\n        labelId=\"labelId1\"\n      />\n      <radio-block\n        label=\"No\"\n        id=\"anyRadioId255\"\n        value=\"false\"\n        name=\"nameChange1\"\n        labelId=\"labelId255\"\n      />\n    </c-radios>\n  </fieldset-block>\n</template>\n<script lang=\"ts\">\nimport Vue from \"vue\";\nimport FieldsetBlock from \"@govtnz/ds/build/vue-ts/FieldsetBlock.vue\";\nimport H1 from \"@govtnz/ds/build/vue-ts/H1.vue\";\nimport Radios from \"@govtnz/ds/build/vue-ts/Radios.vue\";\nimport RadioBlock from \"@govtnz/ds/build/vue-ts/RadioBlock.vue\";\n\nexport default {\n  components: {\n    \"fieldset-block\": FieldsetBlock,\n    \"c-h1\": H1,\n    \"c-radios\": Radios,\n    \"radio-block\": RadioBlock\n  }\n};\n</script>\n",
+    "scss": "",
+    "html": "<!--\nRemember to add these styles:\nin CSS: FieldsetBlock.css, H1.css, Radios.css, RadioBlock.css\nOR in Sass (SCSS): FieldsetBlock.scss, H1.scss, Radios.scss, RadioBlock.scss\n-->\n<div class=\"g-fieldsetBlock-form-group\">\n  <fieldset class=\"g-fieldsetBlock-fieldset\" aria-describedby=\"hintId1\">\n    <legend class=\"g-fieldsetBlock-fieldset__legend\">\n      <h1 class=\"g-heading g-heading-l\" id=\"nameChangeId1\">\n        Have you changed your name?\n      </h1>\n    </legend>\n    <div class=\"g-fieldsetBlock-hint\" id=\"hintId1\">\n      <p>\n        This includes changing your last name or spelling your name differently.\n      </p>\n    </div>\n    <div>\n      <div class=\"g-radios g-radios--inline g-radios--big\">\n        <div class=\"g-radios__item\">\n          <input\n            class=\"g-radios__input\"\n            id=\"anyRadioId\"\n            name=\"nameChange1\"\n            type=\"radio\"\n            value=\"true\"\n          />\n          <label class=\"g-radios__label\" for=\"anyRadioId\"> Yes </label>\n        </div>\n        <div class=\"g-radios__item\">\n          <input\n            class=\"g-radios__input\"\n            id=\"anyRadioId255\"\n            name=\"nameChange1\"\n            type=\"radio\"\n            value=\"false\"\n          />\n          <label class=\"g-radios__label\" for=\"anyRadioId255\"> No </label>\n        </div>\n      </div>\n    </div>\n  </fieldset>\n</div>\n",
+    "css": "",
+    "twig-embed": "{% embed \"FieldsetBlock.html.twig\"  with {'hintId':'hintId1'} only %}{% block legend %}{% embed \"H1.html.twig\"  with {'styleSize':'large', 'id':'nameChangeId1', 'children':'Have you changed your name?'} only %}{% endembed %}{% endblock %}{% block hint %}<p >This includes changing your last name or spelling your name differently.</p>{% endblock %}{% block children %}{% embed \"Radios.html.twig\"  with {'inline':'true', 'big':'true'} only %}{% block children %}{% embed \"RadioBlock.html.twig\"  with {'label':'Yes', 'id':'anyRadioId', 'value':'true', 'name':'nameChange1', 'labelId':'labelId1'} only %}{% endembed %}\n{% embed \"RadioBlock.html.twig\"  with {'label':'No', 'id':'anyRadioId255', 'value':'false', 'name':'nameChange1', 'labelId':'labelId255'} only %}{% endembed %}{% endblock %}{% endembed %}{% endblock %}{% endembed %}",
     "angular": ""
   },
   {
