+++ packages/govtnz-ds-website/static/components/Alerts__example1.html
@@ -33,7 +33,7 @@
       <a href="/">New Zealand Government design system</a>.
     </div>
     <div id="root" class="css-changes-for-example-only">
-      <button type="button" aria-controls="id047212994730638624" aria-expanded="false" class="g-button g-button--secondary">Toggle live <!-- -->info<!-- --> alert</button><div id="id047212994730638624"> <!-- --> <div aria-atomic="true" aria-live="polite" role="note"> <!-- --> </div> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> </div>
+      <button type="button" aria-controls="id0594738543760734" aria-expanded="false" class="g-button g-button--secondary">Toggle live <!-- -->info<!-- --> alert</button><div id="id0594738543760734"> <!-- --> <div aria-atomic="true" aria-live="polite" role="note"> <!-- --> </div> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> </div>
     </div>
     <script>
       function afterRender() {
