+++ packages/govtnz-ds-website/static/components/Alerts__example3.html
@@ -33,7 +33,7 @@
       <a href="/">New Zealand Government design system</a>.
     </div>
     <div id="root" class="css-changes-for-example-only">
-      <button type="button" aria-controls="id0781516153445496" aria-expanded="false" class="g-button g-button--secondary">Toggle live <!-- -->info<!-- --> alert</button><div id="id0781516153445496"> <!-- --> <div aria-atomic="true" aria-live="polite" role="note"> <!-- --> </div> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> </div>
+      <button type="button" aria-controls="id04981856470335446" aria-expanded="false" class="g-button g-button--secondary">Toggle live <!-- -->info<!-- --> alert</button><div id="id04981856470335446"> <!-- --> <div aria-atomic="true" aria-live="polite" role="note"> <!-- --> </div> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> <!-- --> </div>
     </div>
     <script>
       function afterRender() {
