+++ packages/govtnz-ds-website/static/components/Radios__example2.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html lang="en-nz">
   <head>
-    <title>Example: Radio buttons | Design System - Alpha | NZ Government</title>
+    <title>Example: Inline radio buttons | Design System - Alpha | NZ Government</title>
     <meta charset="utf-8" />
     <meta http-equiv="x-ua-compatible" content="ie=edge" />
     <meta
@@ -33,7 +33,7 @@
       <a href="/">New Zealand Government design system</a>.
     </div>
     <div id="root" class="css-changes-for-example-only">
-      <div class="g-fieldsetBlock-form-group"> <fieldset aria-describedby="hintId1" class="g-fieldsetBlock-fieldset"> <legend class="g-fieldsetBlock-fieldset__legend"> <h1 class="g-heading g-heading-l" id="nameChangeId1"> <!-- -->Have you changed your name?<!-- --> </h1> </legend> <!-- --> <div class="g-fieldsetBlock-hint" id="hintId1"> <p>This includes changing your last name or spelling your name differently.</p> </div> <!-- --> <!-- --> <div><div class="g-radios g-radios--inline g-radios--big"> <div class="g-radios__item"> <input type="radio" class="g-radios__input" id="anyRadioId" name="nameChange1" value="true"/> <label class="g-radios__label" for="anyRadioId"> <!-- -->Yes<!-- --> </label> <!-- --> </div><div class="g-radios__item"> <input type="radio" class="g-radios__input" id="anyRadioId255" name="nameChange1" value="false"/> <label class="g-radios__label" for="anyRadioId255"> <!-- -->No<!-- --> </label> <!-- --> </div> </div></div> </fieldset> </div>
+      <div class="g-fieldsetBlock-form-group"> <fieldset aria-describedby="hintId2221" class="g-fieldsetBlock-fieldset"> <legend class="g-fieldsetBlock-fieldset__legend"> <h1 class="g-heading g-heading-l" id="nameChangeId6"> <!-- -->Have you changed your name?<!-- --> </h1> </legend> <!-- --> <div class="g-fieldsetBlock-hint" id="hintId2221"> <p>This includes changing your last name or spelling your name differently.</p> </div> <!-- --> <!-- --> <div><div class="g-radios g-radios--inline"> <div class="g-radios__item"> <input type="radio" class="g-radios__input" id="anyRadioId55" name="nameChange1" value="true"/> <label class="g-radios__label" for="anyRadioId55"> <!-- -->Yes<!-- --> </label> <!-- --> </div><div class="g-radios__item"> <input type="radio" class="g-radios__input" id="anyRadioId2551" name="nameChange1" value="false"/> <label class="g-radios__label" for="anyRadioId2551"> <!-- -->No<!-- --> </label> <!-- --> </div> </div></div> </fieldset> </div>
     </div>
     <script>
       function afterRender() {
