+++ packages/govtnz-ds-website/static/components/Radios__example2.tsx
@@ -13,11 +13,11 @@ const ExampleSection = ({ children }) => (
 );
 const Example = ({ children }) => <Fragment>{children}</Fragment>;
 
-var PageContent = (props) => (<Example title="Example: Radio buttons" {...onChangeGenerator({})}>
-        <FieldsetBlock legend={<H1 styleSize="large" id="nameChangeId1">Have you changed your name?</H1>} hint={<p>This includes changing your last name or spelling your name differently.</p>} hintId="hintId1" {...onChangeGenerator({})}>
-            <Radios inline big {...onChangeGenerator({})}>
-                <RadioBlock label="Yes" id="anyRadioId" value="true" name="nameChange1" labelId="labelId1" {...onChangeGenerator({"value":"true"})}></RadioBlock>
-                <RadioBlock label="No" id="anyRadioId255" value="false" name="nameChange1" labelId="labelId255" {...onChangeGenerator({"value":"false"})}></RadioBlock>
+var PageContent = (props) => (<Example title="Example: Inline radio buttons" {...onChangeGenerator({})}>
+        <FieldsetBlock legend={<H1 styleSize="large" id="nameChangeId6">Have you changed your name?</H1>} hint={<p>This includes changing your last name or spelling your name differently.</p>} hintId="hintId2221" {...onChangeGenerator({})}>
+            <Radios inline {...onChangeGenerator({})}>
+                <RadioBlock label="Yes" id="anyRadioId55" value="true" name="nameChange1" labelId="labelId1" {...onChangeGenerator({"value":"true"})}></RadioBlock>
+                <RadioBlock label="No" id="anyRadioId2551" value="false" name="nameChange1" labelId="labelId2" {...onChangeGenerator({"value":"false"})}></RadioBlock>
             </Radios>
         </FieldsetBlock>
     </Example>);
