+++ packages/govtnz-ds-website/static/components/Radios__example3.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html lang="en-nz">
   <head>
-    <title>Example: Error state #1 (Radio buttons) | Design System - Alpha | NZ Government</title>
+    <title>Example: Radio button items with hints | Design System - Alpha | NZ Government</title>
     <meta charset="utf-8" />
     <meta http-equiv="x-ua-compatible" content="ie=edge" />
     <meta
@@ -13,7 +13,7 @@
     <script src="./Radios__example3.js"></script>
     <link rel="stylesheet" href="../css/Button.css" />
     <link rel="stylesheet" href="../IframePage.css" />
-    <link rel="stylesheet" href="../css/FieldsetBlock.css"><link rel="stylesheet" href="../css/H1.css"><link rel="stylesheet" href="../css/Radios.css"><link rel="stylesheet" href="../css/RadioBlock.css">
+    <link rel="stylesheet" href="../css/FieldsetBlock.css"><link rel="stylesheet" href="../css/H1.css"><link rel="stylesheet" href="../css/RadioBlock.css">
   </head>
   <body>
     <noscript>
@@ -33,7 +33,7 @@
       <a href="/">New Zealand Government design system</a>.
     </div>
     <div id="root" class="css-changes-for-example-only">
-      <div class="g-fieldsetBlock-form-group g-fieldsetBlock-form-group--error"> <fieldset aria-describedby="hintId5 errorId5" class="g-fieldsetBlock-fieldset"> <legend class="g-fieldsetBlock-fieldset__legend"> <h1 class="g-heading g-heading-l" id="nameChangeId5"> <!-- -->Have you changed your name?<!-- --> </h1> </legend> <!-- --> <div class="g-fieldsetBlock-hint" id="hintId5"> <p>This includes changing your last name or spelling your name differently.</p> </div> <!-- --> <!-- --> <div class="g-fieldsetBlock-error-message" id="errorId5"> <span class="g-fieldsetBlock-visually-hidden"> <!-- -->Error:<!-- --> </span> <!-- -->Select &#x27;yes&#x27; if you have changed your name.<!-- --> </div> <!-- --> <div><div class="g-radios g-radios--big"> <div class="g-radios__item"> <input type="radio" class="g-radios__input" id="anyRadioId9234" name="nameChange5" value="true"/> <label class="g-radios__label" for="anyRadioId9234"> <!-- -->Yes<!-- --> </label> <!-- --> </div><div class="g-radios__item"> <input type="radio" class="g-radios__input" id="anyRadioId10" name="nameChange5" value="false"/> <label class="g-radios__label" for="anyRadioId10"> <!-- -->No<!-- --> </label> <!-- --> </div> </div></div> </fieldset> </div>
+      <div class="g-fieldsetBlock-form-group"> <fieldset aria-describedby="hintId" class="g-fieldsetBlock-fieldset"> <legend class="g-fieldsetBlock-fieldset__legend"> <h1 class="g-heading g-heading-l" id="providerChoiceTitle1"> <!-- -->Choose one provider<!-- --> </h1> </legend> <!-- --> <div class="g-fieldsetBlock-hint" id="hintId"> <p>Select one provider.</p> </div> <!-- --> <!-- --> <div><div class="g-radios__item"> <input type="radio" aria-describedby="someHintId7" class="g-radios__input" id="anyRadioId7234" name="providerChoice3" value="provider1"/> <label class="g-radios__label" for="anyRadioId7234"> <!-- -->Sign in with Aardvark services<!-- --> </label> <!-- --> <div class="g-radios__hint" id="someHintId7"> <!-- -->North Island only<!-- --> </div> <!-- --> </div><div class="g-radios__item"> <input type="radio" aria-describedby="someHintId8" class="g-radios__input" id="anyRadioId8234" name="providerChoice3" value="provider2"/> <label class="g-radios__label" for="anyRadioId8234"> <!-- -->Sign in with Bumblebee services<!-- --> </label> <!-- --> <div class="g-radios__hint" id="someHintId8"> <!-- -->North and South Island<!-- --> </div> <!-- --> </div><div class="g-radios__item"> <input type="radio" aria-describedby="someHintId7234" class="g-radios__input" id="anyRadioId9f" name="providerChoice3" value="provider3"/> <label class="g-radios__label" for="anyRadioId9f"> <!-- -->Sign in with Caterpillar company<!-- --> </label> <!-- --> <div class="g-radios__hint" id="someHintId7234"> <!-- -->All new Zealand territories<!-- --> </div> <!-- --> </div></div> </fieldset> </div>
     </div>
     <script>
       function afterRender() {
