+++ packages/govtnz-ds-website/static/components/Radios__example6.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html lang="en-nz">
   <head>
-    <title>Example: Stacked radio buttons | Design System - Alpha | NZ Government</title>
+    <title>Example: Error state #2 (Radio buttons) | Design System - Alpha | NZ Government</title>
     <meta charset="utf-8" />
     <meta http-equiv="x-ua-compatible" content="ie=edge" />
     <meta
@@ -13,7 +13,7 @@
     <script src="./Radios__example6.js"></script>
     <link rel="stylesheet" href="../css/Button.css" />
     <link rel="stylesheet" href="../IframePage.css" />
-    <link rel="stylesheet" href="../css/FieldsetBlock.css"><link rel="stylesheet" href="../css/H1.css"><link rel="stylesheet" href="../css/RadioBlock.css">
+    <link rel="stylesheet" href="../css/FieldsetBlock.css"><link rel="stylesheet" href="../css/H1.css"><link rel="stylesheet" href="../css/Radios.css"><link rel="stylesheet" href="../css/RadioBlock.css">
   </head>
   <body>
     <noscript>
@@ -33,7 +33,7 @@
       <a href="/">New Zealand Government design system</a>.
     </div>
     <div id="root" class="css-changes-for-example-only">
-      <div class="g-fieldsetBlock-form-group"> <fieldset class="g-fieldsetBlock-fieldset"> <legend class="g-fieldsetBlock-fieldset__legend"> <h1 class="g-heading g-heading-l" id="whereLiveTitle"> <!-- -->Where do you live?<!-- --> </h1> </legend> <!-- --> <!-- --> <div><div class="g-radios__item"> <input type="radio" class="g-radios__input" id="anyRadioId3334" name="where" value="north"/> <label class="g-radios__label" for="anyRadioId3334"> <!-- -->North Island<!-- --> </label> <!-- --> </div><div class="g-radios__item"> <input type="radio" class="g-radios__input" id="anyRadioId42344" name="where" value="south"/> <label class="g-radios__label" for="anyRadioId42344"> <!-- -->South Island<!-- --> </label> <!-- --> </div><div class="g-radios__item"> <input type="radio" class="g-radios__input" id="anyRadioId5234" name="where" value="stewart"/> <label class="g-radios__label" for="anyRadioId5234"> <!-- -->Stewart Island<!-- --> </label> <!-- --> </div><div class="g-radios__item"> <input type="radio" class="g-radios__input" id="anyRadioId6234" name="where" value="chatham"/> <label class="g-radios__label" for="anyRadioId6234"> <!-- -->Chatham Islands<!-- --> </label> <!-- --> </div></div> </fieldset> </div>
+      <div class="g-fieldsetBlock-form-group g-fieldsetBlock-form-group--error"> <fieldset aria-describedby="hintId5344 errorId525" class="g-fieldsetBlock-fieldset"> <legend class="g-fieldsetBlock-fieldset__legend"> <h1 class="g-heading g-heading-l" id="nameChangeId8"> <!-- -->Have you changed your name?<!-- --> </h1> </legend> <!-- --> <div class="g-fieldsetBlock-hint" id="hintId5344"> <p>This includes changing your last name or spelling your name differently.</p> </div> <!-- --> <!-- --> <div class="g-fieldsetBlock-error-message" id="errorId525"> <span class="g-fieldsetBlock-visually-hidden"> <!-- -->Error:<!-- --> </span> <!-- -->Select &#x27;yes&#x27; if you have changed your name<!-- --> </div> <!-- --> <div><div class="g-radios g-radios--inline"> <div class="g-radios__item"> <input type="radio" class="g-radios__input" id="anyRadioId9" name="nameChange5" value="true"/> <label class="g-radios__label" for="anyRadioId9"> <!-- -->Yes<!-- --> </label> <!-- --> </div><div class="g-radios__item"> <input type="radio" class="g-radios__input" id="anyRadioId10234" name="nameChange5" value="false"/> <label class="g-radios__label" for="anyRadioId10234"> <!-- -->No<!-- --> </label> <!-- --> </div> </div></div> </fieldset> </div>
     </div>
     <script>
       function afterRender() {
