+++ packages/govtnz-ds-website/static/components/Radios__example6.js
@@ -1,1 +1,1 @@
-!function(e){var t={};function r(n){if(t[n])return t[n].exports;var a=t[n]={i:n,l:!1,exports:{}};return e[n].call(a.exports,a,a.exports,r),a.l=!0,a.exports}r.m=e,r.c=t,r.d=function(e,t,n){r.o(e,t)||Object.defineProperty(e,t,{enumerable:!0,get:n})},r.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},r.t=function(e,t){if(1&t&&(e=r(e)),8&t)return e;if(4&t&&"object"==typeof e&&e&&e.__esModule)return e;var n=Object.create(null);if(r.r(n),Object.defineProperty(n,"default",{enumerable:!0,value:e}),2&t&&"string"!=typeof e)for(var a in e)r.d(n,a,function(t){return e[t]}.bind(null,a));return n},r.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return r.d(t,"a",t),t},r.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},r.p="",r(r.s=5)}([function(e,t){e.exports=React},function(e,t,r){"use strict";var n=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=n(r(0));t.default=function(e){var t=e.id,r=e.hintId,n=e.disabled,l=e.readOnly,o=e.autoFocus,u=e.name,i=e.value,c=e.checked,d=e.onChange,f=e.ref,s=e.label,m=e.hint;return a.default.createElement("div",{className:"g-radios__item"}," ",a.default.createElement("input",{"aria-describedby":r,className:"g-radios__input",id:t,name:u,type:"radio",disabled:void 0!==n?"true"===n.toString():void 0,readOnly:l,autoFocus:o,value:i,checked:c,onChange:d,ref:f})," ",a.default.createElement("label",{className:"g-radios__label",htmlFor:t}," ",void 0!==s?s:a.default.createElement(a.default.Fragment,null,"Label text")," ")," ",void 0!==r?a.default.createElement(a.default.Fragment,null," ",a.default.createElement("div",{className:"g-radios__hint",id:r}," ",void 0!==m?m:a.default.createElement(a.default.Fragment,null,"Hint text")," ")," "):""," ")}},function(e,t){e.exports=ReactDOM},function(e,t,r){"use strict";var n=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=n(r(0));t.default=function(e){var t=e.errorId,r=e.hintId,n=e.legend,l=e.hint,o=e.error,u=e.children;return a.default.createElement("div",{className:"g-fieldsetBlock-form-group"+(t?" g-fieldsetBlock-form-group--error":"")}," ",a.default.createElement("fieldset",{"aria-describedby":void 0!==r||void 0!==t?(r||"")+(t?" "+t:""):void 0,className:"g-fieldsetBlock-fieldset"}," ",a.default.createElement("legend",{className:"g-fieldsetBlock-fieldset__legend"}," ",void 0!==n?n:a.default.createElement(a.default.Fragment,null,"Legend text")," ")," ",void 0!==r?a.default.createElement(a.default.Fragment,null," ",a.default.createElement("div",{className:"g-fieldsetBlock-hint",id:r}," ",void 0!==l?l:a.default.createElement(a.default.Fragment,null,"Hint text")," ")," "):""," ",void 0!==t?a.default.createElement(a.default.Fragment,null," ",a.default.createElement("div",{className:"g-fieldsetBlock-error-message",id:t}," ",a.default.createElement("span",{className:"g-fieldsetBlock-visually-hidden"}," ","Error:"," ")," ",void 0!==o?o:a.default.createElement(a.default.Fragment,null,"Error text")," ")," "):""," ",a.default.createElement("div",null,void 0!==u?u:a.default.createElement(a.default.Fragment,null,"Fieldset contents"))," ")," ")}},function(e,t,r){"use strict";var n=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=n(r(0)),l={xlarge:"g-heading-xl",large:"g-heading-l",medium:"g-heading-m",small:"g-heading-s",xsmall:"g-heading-xs",xxsmall:"g-heading-xxs"};t.default=function(e){var t=e.styleSize,r=e.marginBottom8,n=e.marginBottom0,o=e.id,u=e.children;return a.default.createElement("h1",{className:"g-heading"+(void 0!==l[t]?" "+l[t]:"")+(r?" g-heading-mb-8":"")+(n?" g-heading-mb-0":""),id:o}," ",void 0!==u?u:a.default.createElement(a.default.Fragment,null,"Example text")," ")}},function(e,t,r){"use strict";r.r(t);var n=r(0),a=r.n(n),l=r(2),o=r.n(l),u=r(3),i=r.n(u),c=r(4),d=r.n(c),f=r(1),s=r.n(f);function m(e,t){return function(e){if(Array.isArray(e))return e}(e)||function(e,t){if("undefined"==typeof Symbol||!(Symbol.iterator in Object(e)))return;var r=[],n=!0,a=!1,l=void 0;try{for(var o,u=e[Symbol.iterator]();!(n=(o=u.next()).done)&&(r.push(o.value),!t||r.length!==t);n=!0);}catch(e){a=!0,l=e}finally{try{n||null==u.return||u.return()}finally{if(a)throw l}}return r}(e,t)||function(e,t){if(!e)return;if("string"==typeof e)return p(e,t);var r=Object.prototype.toString.call(e).slice(8,-1);"Object"===r&&e.constructor&&(r=e.constructor.name);if("Map"===r||"Set"===r)return Array.from(e);if("Arguments"===r||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(r))return p(e,t)}(e,t)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}()}function p(e,t){(null==t||t>e.length)&&(t=e.length);for(var r=0,n=new Array(t);r<t;r++)n[r]=e[r];return n}function b(e,t){var r=Object.keys(e);if(Object.getOwnPropertySymbols){var n=Object.getOwnPropertySymbols(e);t&&(n=n.filter((function(t){return Object.getOwnPropertyDescriptor(e,t).enumerable}))),r.push.apply(r,n)}return r}function v(e){for(var t=1;t<arguments.length;t++){var r=null!=arguments[t]?arguments[t]:{};t%2?b(Object(r),!0).forEach((function(t){y(e,t,r[t])})):Object.getOwnPropertyDescriptors?Object.defineProperties(e,Object.getOwnPropertyDescriptors(r)):b(Object(r)).forEach((function(t){Object.defineProperty(e,t,Object.getOwnPropertyDescriptor(r,t))}))}return e}function y(e,t,r){return t in e?Object.defineProperty(e,t,{value:r,enumerable:!0,configurable:!0,writable:!0}):e[t]=r,e}function g(e){return(g="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function h(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}function O(e,t){for(var r=0;r<t.length;r++){var n=t[r];n.enumerable=n.enumerable||!1,n.configurable=!0,"value"in n&&(n.writable=!0),Object.defineProperty(e,n.key,n)}}function E(e,t,r){return t&&O(e.prototype,t),r&&O(e,r),e}function _(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&j(e,t)}function j(e,t){return(j=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}function w(e){var t=function(){if("undefined"==typeof Reflect||!Reflect.construct)return!1;if(Reflect.construct.sham)return!1;if("function"==typeof Proxy)return!0;try{return Date.prototype.toString.call(Reflect.construct(Date,[],(function(){}))),!0}catch(e){return!1}}();return function(){var r,n=P(e);if(t){var a=P(this).constructor;r=Reflect.construct(n,arguments,a)}else r=n.apply(this,arguments);return S(this,r)}}function S(e,t){return!t||"object"!==g(t)&&"function"!=typeof t?x(e):t}function x(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}function P(e){return(P=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function I(){return(I=Object.assign||function(e){for(var t=1;t<arguments.length;t++){var r=arguments[t];for(var n in r)Object.prototype.hasOwnProperty.call(r,n)&&(e[n]=r[n])}return e}).apply(this,arguments)}var F=function(e){var t=e.children;return a.a.createElement(n.Fragment,null,t)},C=function(e){return a.a.createElement(F,I({title:"Example: Stacked radio buttons"},R({})),a.a.createElement(i.a,I({legend:a.a.createElement(d.a,{styleSize:"large",id:"whereLiveTitle"},"Where do you live?")},R({})),a.a.createElement(s.a,I({label:"North Island",id:"anyRadioId3334",name:"where",value:"north",labelId:"labelId3"},R({value:"north"}))),a.a.createElement(s.a,I({label:"South Island",id:"anyRadioId42344",name:"where",value:"south",labelId:"labelId4"},R({value:"south"}))),a.a.createElement(s.a,I({label:"Stewart Island",id:"anyRadioId5234",name:"where",value:"stewart",labelId:"labelId5"},R({value:"stewart"}))),a.a.createElement(s.a,I({label:"Chatham Islands",id:"anyRadioId6234",name:"where",value:"chatham",labelId:"labelId6"},R({value:"chatham"})))))},M=a.a.createContext(void 0),R=(n.Component,a.a.Component,a.a.Component,function(e){var t=m(Object(n.useState)(e&&e.value),2),r=t[0],a=t[1],l=m(Object(n.useState)(e&&e.value2),2),o=l[0],u=l[1],i=m(Object(n.useState)(e&&e.value3),2),c=i[0],d=i[1];return{value:r,value2:o,value3:c,onChange:function(e){return a(e.target.value)},onChange2:function(e){return u(e.target.value)},onChange3:function(e){return d(e.target.value)}}});document.addEventListener("DOMContentLoaded",(function(){var e=document.querySelector("#root");e?o.a.hydrate(a.a.createElement(C,null),e,window.afterRender):console.error("Couldn't find app mount point ","#root")}))}]);
+!function(e){var t={};function n(r){if(t[r])return t[r].exports;var a=t[r]={i:r,l:!1,exports:{}};return e[r].call(a.exports,a,a.exports,n),a.l=!0,a.exports}n.m=e,n.c=t,n.d=function(e,t,r){n.o(e,t)||Object.defineProperty(e,t,{enumerable:!0,get:r})},n.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},n.t=function(e,t){if(1&t&&(e=n(e)),8&t)return e;if(4&t&&"object"==typeof e&&e&&e.__esModule)return e;var r=Object.create(null);if(n.r(r),Object.defineProperty(r,"default",{enumerable:!0,value:e}),2&t&&"string"!=typeof e)for(var a in e)n.d(r,a,function(t){return e[t]}.bind(null,a));return r},n.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return n.d(t,"a",t),t},n.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},n.p="",n(n.s=6)}([function(e,t){e.exports=React},function(e,t,n){"use strict";var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=r(n(0));t.default=function(e){var t=e.id,n=e.hintId,r=e.disabled,o=e.readOnly,l=e.autoFocus,u=e.name,i=e.value,c=e.checked,f=e.onChange,d=e.ref,s=e.label,m=e.hint;return a.default.createElement("div",{className:"g-radios__item"}," ",a.default.createElement("input",{"aria-describedby":n,className:"g-radios__input",id:t,name:u,type:"radio",disabled:void 0!==r?"true"===r.toString():void 0,readOnly:o,autoFocus:l,value:i,checked:c,onChange:f,ref:d})," ",a.default.createElement("label",{className:"g-radios__label",htmlFor:t}," ",void 0!==s?s:a.default.createElement(a.default.Fragment,null,"Label text")," ")," ",void 0!==n?a.default.createElement(a.default.Fragment,null," ",a.default.createElement("div",{className:"g-radios__hint",id:n}," ",void 0!==m?m:a.default.createElement(a.default.Fragment,null,"Hint text")," ")," "):""," ")}},function(e,t){e.exports=ReactDOM},function(e,t,n){"use strict";var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=r(n(0));t.default=function(e){var t=e.errorId,n=e.hintId,r=e.legend,o=e.hint,l=e.error,u=e.children;return a.default.createElement("div",{className:"g-fieldsetBlock-form-group"+(t?" g-fieldsetBlock-form-group--error":"")}," ",a.default.createElement("fieldset",{"aria-describedby":void 0!==n||void 0!==t?(n||"")+(t?" "+t:""):void 0,className:"g-fieldsetBlock-fieldset"}," ",a.default.createElement("legend",{className:"g-fieldsetBlock-fieldset__legend"}," ",void 0!==r?r:a.default.createElement(a.default.Fragment,null,"Legend text")," ")," ",void 0!==n?a.default.createElement(a.default.Fragment,null," ",a.default.createElement("div",{className:"g-fieldsetBlock-hint",id:n}," ",void 0!==o?o:a.default.createElement(a.default.Fragment,null,"Hint text")," ")," "):""," ",void 0!==t?a.default.createElement(a.default.Fragment,null," ",a.default.createElement("div",{className:"g-fieldsetBlock-error-message",id:t}," ",a.default.createElement("span",{className:"g-fieldsetBlock-visually-hidden"}," ","Error:"," ")," ",void 0!==l?l:a.default.createElement(a.default.Fragment,null,"Error text")," ")," "):""," ",a.default.createElement("div",null,void 0!==u?u:a.default.createElement(a.default.Fragment,null,"Fieldset contents"))," ")," ")}},function(e,t,n){"use strict";var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=r(n(0)),o={xlarge:"g-heading-xl",large:"g-heading-l",medium:"g-heading-m",small:"g-heading-s",xsmall:"g-heading-xs",xxsmall:"g-heading-xxs"};t.default=function(e){var t=e.styleSize,n=e.marginBottom8,r=e.marginBottom0,l=e.id,u=e.children;return a.default.createElement("h1",{className:"g-heading"+(void 0!==o[t]?" "+o[t]:"")+(n?" g-heading-mb-8":"")+(r?" g-heading-mb-0":""),id:l}," ",void 0!==u?u:a.default.createElement(a.default.Fragment,null,"Example text")," ")}},function(e,t,n){"use strict";var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=r(n(0));t.default=function(e){var t=e.inline,n=e.big,r=e.children;return a.default.createElement("div",{className:"g-radios"+(t?" g-radios--inline":"")+(n?" g-radios--big":"")}," ",void 0!==r?r:a.default.createElement(a.default.Fragment,null,"Radio components")," ")}},function(e,t,n){"use strict";n.r(t);var r=n(0),a=n.n(r),o=n(2),l=n.n(o),u=n(3),i=n.n(u),c=n(4),f=n.n(c),d=n(5),s=n.n(d),m=n(1),p=n.n(m);function g(e,t){return function(e){if(Array.isArray(e))return e}(e)||function(e,t){if("undefined"==typeof Symbol||!(Symbol.iterator in Object(e)))return;var n=[],r=!0,a=!1,o=void 0;try{for(var l,u=e[Symbol.iterator]();!(r=(l=u.next()).done)&&(n.push(l.value),!t||n.length!==t);r=!0);}catch(e){a=!0,o=e}finally{try{r||null==u.return||u.return()}finally{if(a)throw o}}return n}(e,t)||function(e,t){if(!e)return;if("string"==typeof e)return y(e,t);var n=Object.prototype.toString.call(e).slice(8,-1);"Object"===n&&e.constructor&&(n=e.constructor.name);if("Map"===n||"Set"===n)return Array.from(e);if("Arguments"===n||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n))return y(e,t)}(e,t)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}()}function y(e,t){(null==t||t>e.length)&&(t=e.length);for(var n=0,r=new Array(t);n<t;n++)r[n]=e[n];return r}function v(e,t){var n=Object.keys(e);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(e);t&&(r=r.filter((function(t){return Object.getOwnPropertyDescriptor(e,t).enumerable}))),n.push.apply(n,r)}return n}function b(e){for(var t=1;t<arguments.length;t++){var n=null!=arguments[t]?arguments[t]:{};t%2?v(Object(n),!0).forEach((function(t){h(e,t,n[t])})):Object.getOwnPropertyDescriptors?Object.defineProperties(e,Object.getOwnPropertyDescriptors(n)):v(Object(n)).forEach((function(t){Object.defineProperty(e,t,Object.getOwnPropertyDescriptor(n,t))}))}return e}function h(e,t,n){return t in e?Object.defineProperty(e,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[t]=n,e}function O(e){return(O="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function _(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}function E(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function j(e,t,n){return t&&E(e.prototype,t),n&&E(e,n),e}function S(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&x(e,t)}function x(e,t){return(x=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}function P(e){var t=function(){if("undefined"==typeof Reflect||!Reflect.construct)return!1;if(Reflect.construct.sham)return!1;if("function"==typeof Proxy)return!0;try{return Date.prototype.toString.call(Reflect.construct(Date,[],(function(){}))),!0}catch(e){return!1}}();return function(){var n,r=C(e);if(t){var a=C(this).constructor;n=Reflect.construct(r,arguments,a)}else n=r.apply(this,arguments);return w(this,n)}}function w(e,t){return!t||"object"!==O(t)&&"function"!=typeof t?I(e):t}function I(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}function C(e){return(C=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function F(){return(F=Object.assign||function(e){for(var t=1;t<arguments.length;t++){var n=arguments[t];for(var r in n)Object.prototype.hasOwnProperty.call(n,r)&&(e[r]=n[r])}return e}).apply(this,arguments)}var M=function(e){var t=e.children;return a.a.createElement(r.Fragment,null,t)},N=function(e){return a.a.createElement(M,F({title:"Example: Error state #2 (Radio buttons)"},D({})),a.a.createElement(i.a,F({legend:a.a.createElement(f.a,{styleSize:"large",id:"nameChangeId8"},"Have you changed your name?"),hint:a.a.createElement("p",null,"This includes changing your last name or spelling your name differently."),hintId:"hintId5344",errorId:"errorId525",error:"Select 'yes' if you have changed your name"},D({})),a.a.createElement(s.a,F({inline:!0},D({})),a.a.createElement(p.a,F({label:"Yes",id:"anyRadioId9",value:"true",name:"nameChange5",labelId:"labelId9"},D({value:"true"}))),a.a.createElement(p.a,F({label:"No",id:"anyRadioId10234",value:"false",name:"nameChange5",labelId:"labelId10"},D({value:"false"}))))))},R=a.a.createContext(void 0),D=(r.Component,a.a.Component,a.a.Component,function(e){var t=g(Object(r.useState)(e&&e.value),2),n=t[0],a=t[1],o=g(Object(r.useState)(e&&e.value2),2),l=o[0],u=o[1],i=g(Object(r.useState)(e&&e.value3),2),c=i[0],f=i[1];return{value:n,value2:l,value3:c,onChange:function(e){return a(e.target.value)},onChange2:function(e){return u(e.target.value)},onChange3:function(e){return f(e.target.value)}}});document.addEventListener("DOMContentLoaded",(function(){var e=document.querySelector("#root");e?l.a.hydrate(a.a.createElement(N,null),e,window.afterRender):console.error("Couldn't find app mount point ","#root")}))}]);
